% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extras.R
\name{add_log_axis_ticks}
\alias{add_log_axis_ticks}
\title{Add ticks for a log-axis to a plot.}
\usage{
add_log_axis_ticks(side, lower = 1e-30, upper = 1e+30, ...)
}
\arguments{
\item{side}{integer in 1,2,3,4, giving the side to put the ticks on (see ?axis for details)}

\item{lower}{lower tick limit}

\item{upper}{upper tick limit}

\item{...}{further arguments to \code{axis}}
}
\value{
nothing, but adds ticks to the axis
}
\description{
The tick locations will be similar to
...,0.1,0.2,0.3,...,0.9,1,2,3,4,...,9,10,20,30,40,....
}
\examples{
plot(10**runif(1000), log='y')
# Add the axis ticks -- note we can use optional arguments to 'axis'
add_log_axis_ticks(side=2, lwd.ticks=0.5)

}
