% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{adjust_longitude_by_360_deg}
\alias{adjust_longitude_by_360_deg}
\title{Given lon-lat point(s) p0, and reference point(s), change the notation
of the longitude of p0 so that it differs by < 360 degrees from the
longitude of the reference point}
\usage{
adjust_longitude_by_360_deg(p0, reference_point)
}
\arguments{
\item{p0}{numeric matrix with columns lon/lat, or a vector with a single
point}

\item{reference_point}{same as p0.}
}
\value{
matrix with the same shape and coordinates as p0, but with
longitudes adjusted as required
}
\description{
Given lon-lat point(s) p0, and reference point(s), change the notation
of the longitude of p0 so that it differs by < 360 degrees from the
longitude of the reference point
}

