% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_dispersion_relation.R
\name{airy_period}
\alias{airy_period}
\title{Compute wave period given wavelength and depth}
\usage{
airy_period(lambda, h, g = 9.81, full_output = FALSE)
}
\arguments{
\item{lambda}{wavelength}

\item{h}{depth}

\item{g}{gravity}

\item{full_output}{if true return full optimization info}
}
\value{
the wave period, along with optimization details if full_output=TRUE
}
\description{
Assumes linear (Airy) wave theory
}
\examples{
## See further examples in the help for airy_wavelength
period = airy_period(10000, 100)

}
