% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_dispersion_relation.R
\name{airy_wavelength}
\alias{airy_wavelength}
\title{Compute wavelength given period and depth}
\usage{
airy_wavelength(period, h, g = 9.81, full_output = FALSE)
}
\arguments{
\item{period}{wave period}

\item{h}{depth}

\item{g}{gravity}

\item{full_output}{if true return full optimization info}
}
\value{
the wavelength, along with optimization information if full_output=TRUE
}
\description{
Assumes linear (Airy) wave theory
}
\examples{
   ## Shallow water limit
   wl = 20000
   h = 20
   period0 = airy_period(wl, h)
   wavelength0 = airy_wavelength(period0, h)

   stopifnot(abs(wl - wavelength0) < 1.0e-05*wavelength0)
   stopifnot(abs(period0 - wl/sqrt(9.81 * h)) < 1.0e-05 * period0)

   ## Deep water limit
   wl = 2
   h = 200
   
   period0 = airy_period(wl, h)
   wavelength0 = airy_wavelength(period0, h)

   stopifnot(abs(wl - wavelength0) < 1.0e-05)
   stopifnot(abs(wl - period0**2 * 9.81/(2*pi)) < 1.0e-05 * wl)

   ## Intermediate water limit
   
   wl = 200
   h = 20
   
   period0 = airy_period(wl, h)
   wavelength0 = airy_wavelength(period0, h)

   stopifnot(abs(wl - wavelength0) < 1.0e-05)

   # Check that the dispersion relation is satisfied to within an expected tolerance
   resid = wavelength0 - period0**2 * 9.81 /(2*pi) * tanh( h * 2 * pi / wavelength0)
   stopifnot(abs(resid) < 1.0e-06)

}
