% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{angle_within_dtheta_of_target}
\alias{angle_within_dtheta_of_target}
\title{Test if 'angle' is within 'dtheta' of 'target_angle'}
\usage{
angle_within_dtheta_of_target(angle, target_angle, dtheta)
}
\arguments{
\item{angle}{vector of angles}

\item{target_angle}{vector of target angles}

\item{dtheta}{vector of 'dthetas' or 'angular deviations'}
}
\value{
logical vector with the same length as angle
}
\description{
Function to determine whether the difference between 'angle' and 'target
angle' is less than or equal to 'dtheta'. ALL ANGLES ARE IN DEGREES.
Accounts for circularity of angles.
}
\examples{
   angle = c(340:359, 0:50)
   target_angle = 15
   dtheta = 25
   nearby = angle_within_dtheta_of_target(angle, target_angle, dtheta)

   # Should evaluate TRUE for angles in [350, 0] and [0, 40]
   expected_result = c(rep(FALSE, 10), rep(TRUE, 51), rep(FALSE, 10))
   stopifnot(all(nearby == expected_result))

   # Check that more extreme circularity is ok
   for(dev in c(-360*2, 360*3)){
       angle2 = angle + dev
       nearby2 = angle_within_dtheta_of_target(angle2, target_angle, dtheta)
       stopifnot(all(nearby == expected_result))
   }
   
}
