% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{BilekLay99_magnitude_depth_duration}
\alias{BilekLay99_magnitude_depth_duration}
\title{Data on earthquake magnitude, depth, and rupture duration}
\usage{
BilekLay99_magnitude_depth_duration()
}
\value{
A data.frame with columns 'Mw' (magnitude), 'Depth' (km), 'Duration'
(s), and 'Site'
}
\description{
This function returns magnitude, depth, duration, and site
from the data in the supplementary material of the paper
"Bilek and Lay (1999) Rigidity variations with depth along interplate
megathrust faults in subduction zones, Nature, 400:443-446".
}
