% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_probabilities.R
\name{compute_moment_fraction_from_events_greater_or_equal_than_mwmin}
\alias{compute_moment_fraction_from_events_greater_or_equal_than_mwmin}
\title{Compute the fraction of seismic moment associated with earthquakes having a
magnitude above some threshold}
\usage{
compute_moment_fraction_from_events_greater_or_equal_than_mwmin(
  Mws,
  rate_Mws,
  moment_Mws,
  mwmin
)
}
\arguments{
\item{Mws}{numeric vector with evenly spaced Mw values, increasing, densely
spaced enough for integration}

\item{rate_Mws}{numeric vector with an occurrence rate for each value in
Mws. THIS IS NOT AN EXCEEDANCE RATE, IT IS A RATE FOR EVENTS WITH Mw = Mws[i].
The value will thus depend on how closely spaced Mws is.}

\item{moment_Mws}{vector with the seismic moment for each Mws [follows from
standard formulae]}

\item{mwmin}{numeric We want to know the fraction of the moment associated
with events of this size or larger}
}
\value{
number between 0-1, giving the fraction of moment associated with
events with Mw > mwmin
}
\description{
For long-term slip rate computations we need to know how much slip is caused
by events with Mw > mwmin. This function helps compute that, though it
requires appropriately designed inputs
}
