% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stage_vs_rate_curve_fast.R
\name{convert_Mw_vs_exceedance_rates_2_stage_vs_exceedance_rates}
\alias{convert_Mw_vs_exceedance_rates_2_stage_vs_exceedance_rates}
\title{Get a set of stage-vs-exceedance-rate curves (one for each magnitude-exceedance rate logic-tree branch)}
\usage{

  convert_Mw_vs_exceedance_rates_2_stage_vs_exceedance_rates(logic_tree_rate_curve_Mw,
  logic_tree_rate_curves_exceedance_rate, event_Mw,
  event_conditional_probability, event_max_stage, output_stages)
}
\arguments{
\item{logic_tree_rate_curve_Mw}{vector with magnitudes at which the logic-tree exceedance-rate curves are tabulated}

\item{logic_tree_rate_curves_exceedance_rate}{matrix with one row per logic-tree branch, and one column 
per logic_tree_rate_curve_Mw, giving the exceedance rate at each magnitude.}

\item{event_Mw}{vector of magnitudes for each event. These should be drawn from an evenly spaced set (e.g. 7.2, 7.3, ... 9.8)}

\item{event_conditional_probability}{vector with teh same length as event_Mw giving the conditional probability that 
the event occurs, given that an event with the same Mw occurred.}

\item{event_max_stage}{The max stage for each event at the site of interest}

\item{output_stages}{the stages at which we tabulate outputs}
}
\value{
A matrix with one row for each logic-tree branch, and one column for each output-stage, giving the exceedance-rates
}
\description{
Suppose we have a source zone containing a set of Mw-vs-exceedance rate curves, and a corresponding set
of events, each of which lead to a given max-stage at a particular site. This routine computes a
stage-vs-exceedance-rate curve for each Mw-vs-exceedance rate curve.
}
