% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downdip_3d_lines_on_source.R
\name{create_downdip_lines_on_source_contours}
\alias{create_downdip_lines_on_source_contours}
\title{Make '3D' down-dip lines along source contours [deprecated]}
\usage{
create_downdip_lines_on_source_contours(source_contours,
  desired_subfault_length, contour_depth_attribute, extend_line_fraction,
  orthogonal_near_trench, make_plot)
}
\arguments{
\item{source_contours}{SpatialLinesDataFrame with the source contours and an
attribute 'level' giving the contour depth}

\item{desired_subfault_length}{Desired length of subfaults along-strike}

\item{contour_depth_attribute}{character The name of the column in the
attribute table giving the contour depth}

\item{extend_line_fraction}{To ensure that contour lines intersect downdip
lines at the left/right edges we extend them by this fraction of the
end-to-end source length}

\item{orthogonal_near_trench}{move unit source points along the trench to enhance
orthogonality there. Can reduce numerical artefacts at the trench}

\item{make_plot}{logical (TRUE/FALSE) Plot the result}
}
\value{
a list with the 3d lines
}
\description{
This is as an intermediate step when defining boundaries of
the unit sources. It is not meant to be exported to the main package,
and has been replaced by more advanced routines. Please use
\code{create_downdip_lines_on_source_contours_improved}
}
