% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downdip_3d_lines_on_source.R
\name{create_downdip_lines_on_source_contours_improved}
\alias{create_downdip_lines_on_source_contours_improved}
\title{Downdip lines on source contours}
\usage{
create_downdip_lines_on_source_contours_improved(
  source_contours,
  desired_unit_source_length,
  contour_depth_attribute = "level",
  force_even_point_spacing = FALSE,
  make_plot = FALSE
)
}
\arguments{
\item{source_contours}{SpatialLinesDataFrame with the source contours in lon/lat coordinates}

\item{desired_unit_source_length}{Desired top-edge length of the unit sources in km}

\item{contour_depth_attribute}{Name of attribute in source_contours that contains the depth}

\item{force_even_point_spacing}{Make the lines consist of points that are evenly spaced along each contour.
This will generally lead to low-quality lines, but can be useful for debugging purposes.}

\item{make_plot}{logical. Plot the contours + downdip lines}
}
\value{
a list with the 3d lines
}
\description{
Make downdip lines (lon/lat/depth) which can be used to form the boundaries
for unit sources on source contours
}
