% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{discretized_source_approximate_summary_statistics}
\alias{discretized_source_approximate_summary_statistics}
\title{Compute APPROXIMATE summary statistics for all unit sources in a discretized source
zone}
\usage{
discretized_source_approximate_summary_statistics(discretized_source,
  default_rake = 90, default_slip = 1, make_plot = FALSE,
  depth_in_km = TRUE)
}
\arguments{
\item{discretized_source}{list with an entry 'unit_source_grid' containing a
3 dimensional array defining the vertices of all unit sources for the source
zone (e.g. output of 'discretized_source_from_source_contours')}

\item{default_rake}{numeric degrees, all unit sources are assigned this rake}

\item{default_slip}{numeric m, all unit sources are assigned this slip}

\item{make_plot}{logical, make a plot?}

\item{depth_in_km}{Are depths in km (TRUE) or meters (FALSE)}
}
\value{
data.frame with key summary statistics
}
\description{
The quantities that we output follow those in the i-invall format in the old
URSGA code, although some outputs there (e.g. section number) do not apply
with our algorithms. \cr
The approach used in this routine is faster than discretized_source_summary_statistics,
at the expense of accuracy [the latter draws on the sub-unit-source grid points].
}
\examples{
# Get source contours
puysegur = readOGR(system.file('extdata/puysegur.shp', package='rptha'), layer='puysegur')
# Get downdip lines
puysegur_downdip = readOGR(system.file('extdata/puysegur_downdip.shp', package='rptha'), 
   layer='puysegur_downdip')
# Make discretized_source with 50km x 50km unit-sources (approximately)
puysegur_discretized_source = discretized_source_from_source_contours(
    source_shapefile=puysegur,
   desired_subfault_length=50,
   desired_subfault_width=50,
   downdip_lines=puysegur_downdip)

# Make summary statistics
puysegur_summary_statistics = discretized_source_approximate_summary_statistics(
    puysegur_discretized_source)
print(head(puysegur_summary_statistics))

}

