% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{distance_down_depth}
\alias{distance_down_depth}
\title{Compute distances between 2 points defined by (lon,lat,depth)}
\usage{
distance_down_depth(p1, p2, depth_in_km = TRUE, n = 1000, r = 6378137)
}
\arguments{
\item{p1}{= (lon1, lat1, depth1) and}

\item{p2}{= (lon2, lat2, depth2).}

\item{depth_in_km}{if TRUE depths are in km (otherwise meters)}

\item{n}{= number of intermediate points used in distance computation}

\item{r}{= radius of the earth (m)}
}
\value{
The distance in m
}
\description{
Assume all depths are relative to a regular datum (MSL -- key point is we
ignore topography of earth). \cr
The surface projection of the path is assumed to follow a great circle
The rate of change of depth with respect to surface distance is assumed constant
}

