% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downdip_3d_lines_on_source.R
\name{downdip_lines_to_SpatialLinesDataFrame}
\alias{downdip_lines_to_SpatialLinesDataFrame}
\title{Convert downdip lines to a SpatialLinesDataFrame}
\usage{
downdip_lines_to_SpatialLinesDataFrame(new_xy)
}
\arguments{
\item{new_xy}{output from \code{create_downdip_lines_on_source_contours_improved}}
}
\value{
a SpatialLinesDataFrame with the downdip lines
}
\description{
Convenience function for making a shapefile from the downdip lines. This
is useful for checking the result quality
}
