% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{dtruncGR}
\alias{dtruncGR}
\title{Probability density function for truncated Gutenberg Richter distribution}
\usage{
dtruncGR(x, b, mw_min, mw_max = Inf)
}
\arguments{
\item{x}{vector of quantiles}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}

\item{mw_max}{Maximum mw}
}
\value{
vector with Mw values having non-exceedance probabilities p
}
\description{
Note corresponding functions ptruncGR, qtruncGR, dtruncGR, rtruncGR. See ?rtruncGR
for an example of using the functions for b estimation
}
\examples{
# Get the probability density for some random Mw values
random_Mw = rtruncGR(10, b=0.7, mw_min=6.0, mw_max = 9.5)
random_density = dtruncGR(random_Mw, b=0.7, mw_min=6.0, mw_max=9.5)

}
