% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{fill_polygon_with_grid_points}
\alias{fill_polygon_with_grid_points}
\title{Workhorse function used inside compute_grid_point_areas_in_polygon}
\usage{
fill_polygon_with_grid_points(polygon, new_origin, x_axis_vector, approx_dx,
  approx_dy)
}
\description{
Workhorse function used inside compute_grid_point_areas_in_polygon
}
