% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gArea}
\alias{gArea}
\title{limited replacement for rgeos::gArea using sf functionality}
\usage{
gArea(spgeom, byid = FALSE)
}
\description{
Like rgeos this treats all geometries as Cartesian. Suggest to use
sf::st_area for non-legacy applications.
}
