% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_statistics.R
\name{gauge_zero_crossing_period}
\alias{gauge_zero_crossing_period}
\title{Zero-crossing-period of a gauge time-series}
\usage{
gauge_zero_crossing_period(data_t, data_s, interp_dt = NULL)
}
\arguments{
\item{data_t}{vector of numeric times (usually in seconds)}

\item{data_s}{vector of numeric stages corresponding to data_t (usually in metres)}

\item{interp_dt}{numeric time-step (same units as data_t). We interpolate
the data to a fixed time-step given by interp_dt, before computing the
zero crossing period. If NULL, the minimum spacing in data_t is used.}
}
\value{
The zero crossing period of the gauge data
}
\description{
This function computes both the up-crossing and down-crossing periods and
returns their average.
}
\examples{
x = seq(1,2000, by=2)
period = 30
y = sin(x*2*pi/period)
# Check we can compute this
y_period = gauge_zero_crossing_period(x, y)
stopifnot(abs(y_period/period - 1) < 1.0e-03)

}
