% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gBuffer}
\alias{gBuffer}
\title{limited replacement for rgeos::gBuffer using sf functionality}
\usage{
gBuffer(spgeom, width, quadsegs = 5, byid = FALSE)
}
\description{
limited replacement for rgeos::gBuffer using sf functionality
}
