% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gCentroid}
\alias{gCentroid}
\title{limited replacement for rgeos::gCentroid using sf functionality}
\usage{
gCentroid(spgeom, byid = FALSE)
}
\description{
limited replacement for rgeos::gCentroid using sf functionality
}
