% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gCovers}
\alias{gCovers}
\title{limited replacement for rgeos::gCovers using sf functionality}
\usage{
gCovers(spgeom1, spgeom2 = NULL, byid = FALSE, returnDense = TRUE)
}
\description{
limited replacement for rgeos::gCovers using sf functionality
}
