% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gDistance}
\alias{gDistance}
\title{limited replacement for rgeos::gDistance using sf functionality}
\usage{
gDistance(spgeom1, spgeom2 = NULL, byid = FALSE)
}
\description{
Like rgeos, this treats all datasets as Cartesian. Use sf::st_distance
for non-legacy applications.
}
