% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_events.R
\name{get_all_earthquake_events_of_magnitude_Mw}
\alias{get_all_earthquake_events_of_magnitude_Mw}
\title{Given the summary statistics for a source zone, compute **all** uniform slip events
with moment magnitude Mw.}
\usage{
get_all_earthquake_events_of_magnitude_Mw(Mw, unit_source_stats, mu = 3e+10,
  constant = 9.05)
}
\arguments{
\item{Mw}{Moment magnitude.}

\item{unit_source_stats}{Output of discretized_source_approximate_summary_statistics or similar.}

\item{mu}{Shear modulus in Pascals.}

\item{constant}{value of constant passed to \code{M0_2_Mw}}
}
\value{
A list with information which can be used to create all events.
}
\description{
The events all have the same number of unit sources along-strike and
down-dip. \code{Mw_2_rupture_size} is used to compute the target event
dimensions.
}
