% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{get_depth_dip_at_unit_source_interior_points}
\alias{get_depth_dip_at_unit_source_interior_points}
\title{Interpolate depth and dip at interior points in the unit source [deprecated]}
\usage{
get_depth_dip_at_unit_source_interior_points(unit_source, grid_points,
  fine_downdip_transects, grid_points_strike)
}
\arguments{
\item{unit_source}{matrix with 4 rows and 3 columns (x,y,depth) defining the
unit source boundary. The 4 points should be ordered so that unit_source[1,]
and unit_source[4,] are the top edge (shallower), and unit_source[2,] is on
the same side of the source as unit_source[1,] (so plotted as a line it looks
sensible). The unit source x,y,depth should be in m (local cartesian coords)}

\item{grid_points}{data.frame with x,y coordinates of the points inside the
unit source where we want to find depth/dip/etc}

\item{fine_downdip_transects}{array of dimension (N, 3, 2).
fine_downdip_transects[,,1] contains a line (x,y, depth) which covers the
edge unit_source[1:2,], while fine_downdip_transects[,,2] contains a similar
line which covers the edge unit_source[4:3,]. The x,y,depth coordinates should be in
m in local cartesian coordinates}

\item{grid_points_strike}{The strike for each grid point}
}
\value{
matrix with x, y, depth, dip,.. etc at points inside the unit source
}
\description{
Interpolate depth and dip at interior points in the unit source [deprecated]
}
