% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_simulate_earthquake.R
\name{get_random_seed}
\alias{get_random_seed}
\title{Return the current value of .Random.seed}
\usage{
get_random_seed()
}
\description{
Utility function to get the random seed -- and make it first if required
}
\examples{
 x = get_random_seed()

}

