% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{get_shallow_unit_source_top_edge_strikes}
\alias{get_shallow_unit_source_top_edge_strikes}
\title{Extract the lon-lat midpoints of the top edge of all unit sources
along the 'shallowest' edge of a discrete source (i.e. near the trench), and
compute the strike}
\usage{
get_shallow_unit_source_top_edge_strikes(discretized_source)
}
\arguments{
\item{discretized_source}{A list containing discretized source information
(output of discretized_source_from_source_contours or similar)}
}
\value{
matrix with 3 columns giving the lon,lat,strike of the top edge of
each unit source
}
\description{
Extract the lon-lat midpoints of the top edge of all unit sources
along the 'shallowest' edge of a discrete source (i.e. near the trench), and
compute the strike
}

