% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_events.R
\name{get_unit_source_indices_in_event}
\alias{get_unit_source_indices_in_event}
\title{Find unit sources contained in event}
\usage{
get_unit_source_indices_in_event(earthquake_event, also_return_slip = FALSE)
}
\arguments{
\item{earthquake_event}{A single row of the earthquake event table made with
\code{get_all_earthquake_events}}

\item{also_return_slip}{logical - if TRUE the result is a list with 'inds'
and 'slip', otherwise we directly return 'inds'}
}
\value{
Integer indices in earthquake_event$event_index_string. These correspond
to rows in the unit_source_summary_statistics for the source.
}
\description{
The earthquake event table (generated by \code{get_all_earthquake_events})
contains a column 'event_index_string' containing the indices of
unit sources which contribute to the event, separated by '-'.
This convenience function converts those indices to a numeric vector of integers.
The latter correspond to the row indices of unit_source_statistics (which
is an argument to \code{get_all_earthquake_events}, see the help of that function
for details)
}
