% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gIntersection}
\alias{gIntersection}
\title{limited replacement for rgeos::gIntersection using sf functionality}
\usage{
gIntersection(spgeom1, spgeom2, byid = FALSE, drop_lower_td = FALSE)
}
\description{
Beware this can lead to different ordering of the geometries as compared to
rgeos::gIntersection. Suggest to use sf::st_intersection for non-legacy applications.
}
