% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gUnion}
\alias{gUnion}
\title{limited replacement for rgeos::gUnion using sf functionality}
\usage{
gUnion(spgeom1, spgeom2, byid = FALSE)
}
\description{
limited replacement for rgeos::gUnion using sf functionality
}
