% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{interpolate_3D_path}
\alias{interpolate_3D_path}
\title{Given a 3D path of points (lon,lat,depth), interpolate a new path with n
evenly spaced points (i.e. equal distances between points in 3D)}
\usage{
interpolate_3D_path(threeD_path, n = 100, depth_in_km = TRUE, ndense = (n
  * 10))
}
\arguments{
\item{threeD_path}{matrix of lon,lat,depth defining the threeD path}

\item{n}{desired number of points on the output path}

\item{depth_in_km}{are depths in km? If FALSE, assume depths are in metres}

\item{ndense}{number of points used to support the interpolation (just ensure it is high enough)}
}
\value{
a threeD path with the required number of points
}
\description{
Given a 3D path of points (lon,lat,depth), interpolate a new path with n
evenly spaced points (i.e. equal distances between points in 3D)
}

