% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{interpolate_gc_path}
\alias{interpolate_gc_path}
\title{Interpolate points along a great circle}
\usage{
interpolate_gc_path(surface_path, n = 50)
}
\arguments{
\item{surface_path}{matrix with dim=c(2,3), with columns = (lon, lat, depth)}

\item{n}{number of points in the interior of the path}
}
\value{
matrix with dim=c(n+2, 3) with columns = (lon,lat,depth), which
        interpolate between the 2 points of surface_path
}
\description{
Given a path defined by 2 points, interpolate along the great-circle
and return a matrix with the points along the great circle (where lon is
adjusted for continuity). There will be n+2 points (so n intermediate points)
}

