% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{intersect_surface_path_with_depth_contours}
\alias{intersect_surface_path_with_depth_contours}
\title{Get points along a path which intersect with depth contours}
\usage{
intersect_surface_path_with_depth_contours(surface_path, depth_contours,
  n = 200, contour_depth_attribute = "level", extend_line_fraction = 0)
}
\arguments{
\item{surface_path}{2 points defining a great-circle path (2x2 matrix with
columns lon,lat)}

\item{depth_contours}{depth_contours as SpatialLinesDataFrame}

\item{n}{number of points to used to interpolate along the surface path,
before we find the contours. Note: The interpolation accounts for the
great-circle geometry, while the intersection treats coordinates as
cartesian.  So we need n sufficiently dense that the computed intersection is
nonetheless reasonable.}

\item{contour_depth_attribute}{name of the depth_contours attribute giving
the depth}

\item{extend_line_fraction}{Extend contours at the edges of the source zone
by (approx) this fraction of their length. This can help ensure
intersections at the edges.}
}
\value{
A 3 column matrix with lon/lat/depth giving the intersection with
the contours
}
\description{
Given a line on the earths surface (defined by 2 points, assumed to follow a
great circle path) and some intersecting depth contours, assign depths to
path points using the contours. \cr
Assumes that the depth is changing at a constant rate with respect to the
along-surface distance.
}

