% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kajiura_convolution.R
\name{kajiura_convolution}
\alias{kajiura_convolution}
\title{Call the fortran kajiura_convolution routine via C}
\usage{
kajiura_convolution(depth_inv, initial_deformation_padded, filterXYr,
  kajiuraGmax, lfx, lfy, lnx, lny, output)
}
\arguments{
\item{depth_inv}{matrix with inverse of depth (no NaN's)}

\item{initial_deformation_padded}{matrix with initial deformations, padded
with rows and columns so the filter can be applied to all internal points (dimension lny+2lfy, lnx + 2lfx)}

\item{filterXYr}{matrix with dimensions of filter, containing distances from the central point}

\item{kajiuraGmax}{maximum input value to the kajiuraG function}

\item{lfx}{(see lfy)}

\item{lfy}{filterXYr has dimension = c(lfy, lfx)}

\item{lnx}{(see lny)}

\item{lny}{depth_inv has dimensions = c(lny, lnx)}

\item{output}{matrix with dimension = c(lny, lnx). Output goes here.}
}
\value{
Nothing but modify outputs
}
\description{
Internal routine to speed up kajiura filter. Tried to efficiently
implement the inner loop for the kajiura filter. See the code in
kajuria_filter to understand in detail the arguments.
}
