% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kajiura_filter.R
\name{kajiura_g}
\alias{kajiura_g}
\title{Compute Kajiura's filter function}
\usage{
kajiura_g(r, nt = 1000, verbose = FALSE, recursive_stop_factor = 1e-05)
}
\arguments{
\item{r}{numeric. Value of r in the equation}

\item{nt}{Integer. Number of terms. We compare the sum of the first nt terms, and the
next nt terms, and if they differ by less than 'recursive_stop_factor' we
return the combined sum. Otherwise we double nt and try again.}

\item{verbose}{logical. Print when recurisng}

\item{recursive_stop_factor}{numeric. See information on 'nt' above}
}
\value{
The value of the function at r
}
\description{
The function value is the sum over n in [0, Infty] of
[ 1/pi*sum( (-1)^n*(2*n+1)/( (2*n+1)^2 + r^2)^(3/2)) ].
For a more efficient and vectorized implementation, see kajiura_g_empirical
which approximates this function using splines and a log transform.
}

