% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kajiura_filter.R
\name{kajiura_g_empirical}
\alias{kajiura_g_empirical}
\title{Get a function to efficiently evaluate Kajiura's G function}
\usage{
kajiura_g_empirical(rMax = 9, n = 81)
}
\arguments{
\item{rMax}{maximum value of r at which lookup is desired}

\item{n}{number of points used to empirically approximate the log-function
values.}
}
\value{
Function to compute kajiura_G from an input vector of R values <= rMax.
}
\description{
This function returns a function which can be used to efficiently
evaluate Kajiura's G function. It is computed by tabulating the function
at 'n' points between r = 0 and r=rMax, and then fitting a spline to the
log-transformed result. The latter spline is used to lookup log(kajiura_g), 
and the exponential transform is applied to correct the result. Our tests
suggest the approach is highly accurate, and in R it is much more efficient than
direct computation.
}
