% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kajiura_filter.R
\name{kajiura_smooth_raster}
\alias{kajiura_smooth_raster}
\title{Apply kajiura to a deformation raster}
\usage{
kajiura_smooth_raster(source_raster, new_origin, elevation_raster,
  kj_filter_grid_dxdy = 1000, kj_filter_def_threshold = 0.001,
  kj_cartesian_buffer = 10000, minimum_kj_depth = 10,
  elevation_extraction_x_offset = 0, spherical_input = NULL)
}
\arguments{
\item{source_raster}{the raster (object of class RasterLayer) with the stage
deformation. The background level is assumed to be 0 = MSL}

\item{new_origin}{origin for cartesian coordinate system in which kj filter
is applied (if spherical_input = TRUE)}

\item{elevation_raster}{name of file giving elevation, or the elevation
raster as a RasterLayer object. Negative = below MSL.  Depths for kajiura
filter are taken from this raster, but limited to minimum_kj_depth.}

\item{kj_filter_grid_dxdy}{value for grid_dx and grid_dy in \code{kajiura_filter}. Can
either be a constant or a vector of length 2 (in which case dx is first and dy is second)}

\item{kj_filter_def_threshold}{real (m). Only apply kajiura filter in a
region where the abs(source_raster) exceeds kj_filter_def_threshold. Actually the
region is somewhat larger to avoid edge effects, see kj_cartesian_buffer.}

\item{kj_cartesian_buffer}{real (m). Exand the region where kj_filter is
applied by this distance to avoid boundary effects.}

\item{minimum_kj_depth}{real (m). Minimum possible depth passed to kj
filter (limit on the depth value inferred from the elevation_raster)}

\item{elevation_extraction_x_offset}{Add this number to the 'longitude'
or x coordinate of the source_raster, prior to extracting associated elevation
values from the elevation_raster. The main use is to set it to +360 or -360 to
deal with recentering in lon/lat rasters}

\item{spherical_input}{TRUE/FALSE. Is input raster in lon/lat coordinates.
If NULL, then use isLonLat to determine this}
}
\value{
a raster (source raster with kajiura filter applied)
}
\description{
Kajiura filter applied directly to a raster representing the stage deformation
}
