% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{lonlat2utm}
\alias{lonlat2utm}
\title{Find the utm zone of a lon/lat point}
\usage{
lonlat2utm(lonlat)
}
\arguments{
\item{lonlat}{vector with c(longitude, latitude) in decimal degrees}
}
\value{
proj4string for the UTM Zone
}
\description{
Given a longitude-latitude point in decimal degrees, return its UTM zone
with the WGS84 ellipsoid + datum as a proj4string \cr
Zone number derived based on discussions on stack-exchange and
stack-overflow:
-- http://gis.stackexchange.com/questions/13291/computing-utm-zone-from-lat-long-point
-- http://stackoverflow.com/questions/9186496/determining-utm-zone-to-convert-from-longitude-latitude
}

