% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{lonlat2xyz}
\alias{lonlat2xyz}
\title{Convert spherical lon/lat coordinates to 3D cartesian x,y,z coordinates}
\usage{
lonlat2xyz(p0, r = 6378137)
}
\arguments{
\item{p0}{2 column matrix with columns lon,lat}

\item{r}{radius of sphere (default = earth radius in m)}
}
\value{
matrix with 3 columns x,y,z giving the cartesian coordinates
}
\description{
Convert spherical lon/lat coordinates to 3D cartesian x,y,z coordinates
}

