% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rupture_scaling.R
\name{M0_2_Mw}
\alias{M0_2_Mw}
\title{Compute Moment Magnitude Mw from Seismic Moment M0, and vice-versa}
\usage{
M0_2_Mw(M0, inverse = FALSE)
}
\arguments{
\item{M0}{Seismic Moment (units of Nm) if inverse = FALSE, else Moment Magnitude}

\item{inverse.}{If FALSE, return Mw given M0. If TRUE, return M0 given Mw}
}
\value{
Mw
}
\description{
We use the equation Mw = 2/3 * (log10(M0) - 9.05), see Hanks and Kanamori
(1977) and Bird and Kagan (2004).  Note that slightly different values of the
final constant are sometimes used.
}

