% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_interpolator.R
\name{make_contour_interpolator}
\alias{make_contour_interpolator}
\title{Interpolate along downdip lines defining (cartesian) source contours}
\usage{
make_contour_interpolator(mid_line_with_cutpoints,
  convert_to_cartesian = FALSE, origin_lonlat = NULL, r = 6378137)
}
\arguments{
\item{mid_line_with_cutpoints}{A list of matrices which is typically obtained
from the 'mid_line_with_cutpoints' entry of the output of
\code{discretized_source_from_source_contours}, or directly from
\code{create_downdip_lines_on_source_contours_improved}. All matrices have
the same dimensions (and 3 columns). Each matrix defines a down-dip
line along the source contours (x,y,depth). The depth values are increasing,
and have the same values in each matrix. The list is ordered along strike.}

\item{convert_to_cartesian}{Logical. If TRUE, the mid_line_with_cutpoints
is in spherical lon/lat coordinates, and the return function will account
for this by transforming x,y data to cartesian prior to fitting, using the
provided values of origin_lonlat and r (see
\code{spherical_to_cartesian2d_coordinates}).}

\item{origin_lonlat}{If convert_to_cartesian = TRUE, then this value of
origin_lonlat is passed to \code{spherical_to_cartesian2d_coordinates}).}

\item{r}{radius of the earth. If convert_to_cartesian = TRUE, then this value
of origin_lonlat is passed to \code{spherical_to_cartesian2d_coordinates}).}
}
\value{
a function f(xy, na_buffer_width=0, xy_perturbation_m=NULL) 
which can interpolate depth values along the source contours. Non-zero buffer
widths can be used to allow the function to be applied to points slightly
outside the polygons. xy_perturbation_m can be a matrix of the same shape as xy,
in which case it is added to xy before depth computation. This turns out to be convenient
for computing dip.
}
\description{
Suppose we have a list of matrices defining a grid of (x,y_depth) points over
the source contours, with the along-strike edges aligned with fixed contour
levels. This input can be obtained from the 'mid_line_with_cutpoints' entry
of the output of \code{discretized_source_from_source_contours}, or from the
function \code{create_downdip_lines_on_source_contours_improved}. \cr
This function is used to make a function, which can assign a depth value to
arbitrary x,y points inside the source. Note it can handle conversion from spherical
to cartesian coordinates (with convert_to_cartesian=TRUE)
}

