% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{make_edge_approxfun}
\alias{make_edge_approxfun}
\title{Utility to assist with interpolating depth at points inside a unit source}
\usage{
make_edge_approxfun(unit_source_edge, fine_downdip_transect)
}
\arguments{
\item{unit_source_edge}{matrix with 2 rows and x,y,depth columns, giving the
coordiantes of an edge of the unit source. x,y,depth MUST be local cartesian coordinates (m)}

\item{fine_downdip_transect}{matrix with many rows and x,y,depth columns
which covers the unit source edge, and gives detail to the depth variation.
Must be in local cartesian coordinates, with x, y in m, and depth in km}
}
\value{
a function f(alpha) --> (x, y, depth), where alpha is a coordinate
in [0,1] giving the fraction distance along that part of the
fine_downdip_transect that covers the unit source edge (might be slightly
larger than the unit source edge)
}
\description{
Given one down-dip edge of a unit source, and an overlapping
fine_downdip_transect , make a function f(alpha): [0-1] --> (x,y,depth) which
gives the cartesian coordinates and depth at fraction 'alpha' of the distance
down the unit source edge.
}

