% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downdip_3d_lines_on_source.R
\name{make_source_contours_interpolation_function_list}
\alias{make_source_contours_interpolation_function_list}
\title{Functions to interpolate points on source contours}
\usage{
make_source_contours_interpolation_function_list(
  source_contours,
  contour_depth_attribute = "level"
)
}
\arguments{
\item{source_contours}{SpatialLinesDataFrame containing the contour
lines (in lon/lat coordinates) and an attribute giving the depth.}

\item{contour_depth_attribute}{Name of the attribute giving the depth.}
}
\value{
A list of functions (one for each line). See
\code{make_line_interpolation_fun} for information on what they do. The list
is named based on the corresponding contour depth
}
\description{
Given source contours in lon/lat coordinates, make a list of functions which
map [0,1] to points on each contour line using
\code{make_line_interpolation_fun}. The list has one such function for each
contour line, and is ordered by increasing depth.
}
