% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{make_tsunami_unit_source}
\alias{make_tsunami_unit_source}
\title{Function to make a tsunami for the ij unit source in a discrete source.}
\usage{
make_tsunami_unit_source(i, j, discrete_source, tsunami_surface_points_lonlat,
  approx_dx = NULL, approx_dy = NULL, scale_dxdy = 1,
  depths_in_km = TRUE,
  tsunami_function = unit_source_cartesian_to_okada_tsunami_source, ...)
}
\arguments{
\item{i}{integer. The down-dip index of the unit source}

\item{j}{integer. The along-strike index of the unit source}

\item{discrete_source}{List with discrete source information}

\item{tsunami_surface_points_lonlat}{matrix with lon/lat coordinates of
points where tsunami initial condition will be evaluated.}

\item{approx_dx}{Approximate x spacing of integration points inside unit
source. See unit_source_interior_points_cartesian.}

\item{approx_dy}{Approximate x spacing of integration points inside unit
source. See unit_source_interior_points_cartesian.}

\item{scale_dxdy}{Adjustment factor for approx_dx and approx_dy}

\item{depths_in_km}{logical. Are input depths in the discrete source in km
(TRUE) or m (FALSE).}

\item{tsunami_function}{Function used to make the tunami source from
a cartesian unit source with interior points}

\item{...}{further arguments to tsunami function}
}
\value{
a list with the cartesian unit source, tsunami source, i,j indices
and tsunami surface points
}
\description{
Function to make a tsunami for the ij unit source in a discrete source.
}

