% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{mean_angle}
\alias{mean_angle}
\title{Compute the mean of a set of angles.}
\usage{
mean_angle(angles, degrees = TRUE, method = "complex-mean",
  weights = 1)
}
\arguments{
\item{angles}{numeric vector. Contains angles to compute mean for}

\item{degrees}{logical. Are input angles in degrees (TRUE) or radians (FALSE)}

\item{method}{character.  'complex-mean' gives Arg( mean(exp(i * angles))).}

\item{weights}{Optionally provide a vector with the same length of 'angles'
giving weights for weighted mean.}
}
\description{
This accounts for the fact that e.g. mean(c(1, 359)) should be 0 for degrees
(not 180 as would be true for the standard mean). The 'complex-mean'
method seems to be the most widely used. Note that with this method
the [mean of c(0,0,90)] != 30.
}
\examples{
  # The counter-intuitive case
  m1 = mean_angle(c(0,0,90))
  stopifnot(m1 < 30)
  # A more typical case
  m2 = mean_angle(c(120, 60))
  stopifnot(abs(m2 - 90) < sqrt(.Machine$double.eps))

}
