% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downdip_3d_lines_on_source.R
\name{mid_line_with_cutpoints_from_downdip_sldf_and_source_contours}
\alias{mid_line_with_cutpoints_from_downdip_sldf_and_source_contours}
\title{Make mid_line_with_cutpoints from two shapefiles}
\usage{
mid_line_with_cutpoints_from_downdip_sldf_and_source_contours(
  source_contours,
  downdip_lines,
  contour_depth_attribute = "level",
  buffer_width = 1e-06,
  make_plot = FALSE
)
}
\arguments{
\item{source_contours}{SpatialLinesDataFrame with the source contours. Must have
an attribute giving the depth, with attribute name = contour_depth_attribute}

\item{downdip_lines}{SpatialLinesDataFrame with the downdip_lines. Must have only
one attribute, which is a number giving the order of the line in the along-strike
direction. [e.g. 1, 2, 3, 4 ...]}

\item{contour_depth_attribute}{Name of attribute with contour depth info in source_contours}

\item{buffer_width}{When doing intersections, buffer by this much to avoid round-off.
This should generally be > 0 unless no intersections occur at line end-points.}

\item{make_plot}{Logical. Make a plot of the source contours and downdip lines}
}
\value{
A list of matrices giving the x,y,depth locations of the mid_line_with_cutpoints,
ordered along strike.
}
\description{
Make mid_line_with_cutpoints from the intersection of source_contours with a user-provided
shapefile consisting of down-dip lines which intersect the contours. The latter lines will
define the along-strike boundaries of the unit sources (the down-dip boundaries will be 
created automatically to make the down-dip width fairly even). \cr
While there are other ways to make mid_line_with_cutpoints, this approach can
be advantageous because the user can edit the downdip_lines shapefile to make
sure the lines are nice -- and thus avoid using the numerical-optimization
based methods (which might not always work well, or might need manual
tweaking, or might be slow to apply repeatedly, etc).
}
