% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_mux.R
\docType{class}
\name{mux2data-class}
\alias{mux2data}
\alias{mux2data-class}
\title{Read mux2 data format}
\arguments{
\item{mux2file}{character. Filename of a mux2 file which is to be read}
}
\value{
An object of class mux2data with slots mux2file, loc, t, wave, wave_tail which
hold the data in the file
}
\description{
S4 class to hold data from mux2 files (an output format of the URS tsunami propagation solver)
}
\examples{
\dontrun{
# Hypothetical system file
filename = "mux2_data_file.mux2"
# Read into an object 'x', access with e.g. x@wave, x@t, etc
x = mux2data(filename)
slotNames(x)
# Example of accessing the table holding the point location information
summary(x@loc)
}
}

