% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rupture_scaling.R
\name{Mw_2_rupture_size}
\alias{Mw_2_rupture_size}
\title{Compute earthquake rupture area, width and length from Mw based on an
empirical scaling relation}
\usage{
Mw_2_rupture_size(Mw, relation = "Strasser", detailed = FALSE, CI_sd = 1)
}
\arguments{
\item{Mw}{Moment Magnitude}

\item{relation}{Name for the scaling relation ('Strasser' uses the subduction
interface event relation for Strasser et al 2010.)}

\item{detailed}{logical. If False return a vector with area/width/length,
otherwise provide a list with the latter as well as information on
log10-standard-deviations}

\item{CI_sd}{Logical. If detailed = TRUE, the output includes a positive
and negative confidence interval threshold, both of which are CI_sd
standard deviations away from the mean (in log space where the regression is
computed)}
}
\value{
A numeric vector with the area/width/length (if detailed = FALSE), otherwise
a list with the rupture size statistics as well as upper and lower bounds
for a confidence interval, and information on the log10 standard deviation
of each variable (which can be used to compute any other confidence interval).
}
\description{
Units are km and km^2
}
\examples{
rupture_statistics1 = Mw_2_rupture_size(9.0)
rupture_statistics2 = Mw_2_rupture_size(9.0, detailed=TRUE)
}

