% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_probabilities.R
\name{Mw_exceedance_rate_characteristic_gutenberg_richter}
\alias{Mw_exceedance_rate_characteristic_gutenberg_richter}
\title{Evaluate the rate of earthquakes with magnitude > Mw using a characteristic
variant of the Gutenberg Richter distribution.}
\usage{
Mw_exceedance_rate_characteristic_gutenberg_richter(Mw, a, b, Mw_min, Mw_max)
}
\arguments{
\item{Mw}{Moment magnitude}

\item{a}{The a parameter}

\item{b}{The b parameter}

\item{Mw_min}{the lower truncated moment magnitude}

\item{Mw_max}{the upper truncated moment magnitude}
}
\value{
The rate of events with magnitude > Mw
}
\description{
The 'pure' Gutenberg Richter distribution gives the number of earthquakes
with magnitude >= Mw as:\cr
N_{GR}(x >= Mw) = 10^(a-bMw) \cr
where a and b are constants. Note 10^(a) is the rate of earthquakes with Mw
>= 0. \cr
The characteristic formulation used here ajusts this as: \cr
N_{GR}(x >= Mw) = 10^(a-bMw) if Mw_min <= Mw <= Mw_max \cr
~~~~~~~~~~~~~  = 10^(a-bMw_min) if Mw < Mw_min \cr
~~~~~~~~~~~~~  = 0 if Mw > Mw_max \cr
This means that the rate of earthquakes of size EXACTLY Mw_max is finite.
}
