% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/okada_tsunami.R
\name{okada_tsunami}
\alias{okada_tsunami}
\title{Compute ground surface displacements in linear elastic half-space from known
subfault ruptures, using Okada's (1985) solution for rectangular subfaults}
\usage{
okada_tsunami(elon, elat, edep, strk, dip, lnth, wdt, disl1, disl2, rlon, rlat,
  dstmx = 9e+12, verbose = FALSE)
}
\arguments{
\item{elon}{-- numeric vector with x rupture centroid location (m)}

\item{elat}{-- numeric vector with y of rupture centroid location (m)}

\item{edep}{-- numeric vector with centroid depth of rupture (km)}

\item{strk}{-- numeric vector with strike of sub-fault (degrees clockwise from North)}

\item{dip}{-- numeric vector with dip of rupture (degrees below the horizontal, dipping to the right when looking in the along-strike direction)}

\item{lnth}{-- numeric vector with length of sub-fault (km)}

\item{wdt}{-- numeric vector with width of sub-fault (km)}

\item{disl1}{-- numeric vector with along-strike disloacation on the sub-fault (m)}

\item{disl2}{-- numeric vector with up-dip disloacation on the sub-fault (m)}

\item{rlon}{-- numeric vector with x locations where output is desired (m)}

\item{rlat}{-- numeric vector with y locations where output is desired (m)}

\item{dstmx}{-- optional maximum distance at which sub-faults can cause displacement}

\item{verbose}{-- TRUE/FALSE -- print info on ground deformation}
}
\value{
-- list with edsp, ndsp, zdsp giving the displacements in the
east,north, and vertical directions. Often we are only interested in the
latter.
}
\description{
Complex ruptures can be treated using multiple sub-faults
}
\examples{
# Simple example -- pure thrust fault
mypts=expand.grid(seq(-100.5,100.5,len=100), seq(-100.5,100.5,len=100))*1000
ff=okada_tsunami(0,0,30,0,15,1,1,0,1,
                mypts[,1],mypts[,2])
plot(mypts[,1], mypts[,2],col=heat.colors(10)[cut(ff[[3]],10)],pch=19,asp=1)

## More complex example
ff=okada_tsunami(c(0,0),c(0,50000.),c(30,25),c(0,0),c(15,12),c(1,1),c(1,1),c(0,0),c(1,1),
                 mypts[,1],mypts[,2])

plot(mypts[,1], mypts[,2],col=heat.colors(10)[cut(ff[[3]],10)],pch=19,asp=1)
}

