% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{orthogonal_near_trench}
\alias{orthogonal_near_trench}
\title{Enhance near-trench orthogonality on discretized source [deprecated]}
\usage{
orthogonal_near_trench(top_line, second_line)
}
\arguments{
\item{top_line}{matrix with 2 rows and n columns. First row is longitude,
second row is latitude}

\item{second_line}{as top_line, but for the second line}
}
\value{
an alternative top_line which should have better orthogonality
}
\description{
Suppose top-line is a set of lon/lat points defining the top of a grid of
unit sources, and second line is a set of lon/lat points defining the bottom
of the unit sources. If top-line is at the trench, then numerically it can be
desirable for the unit sources to be highly orthogonal where they intersect
the trench. One way to do this is by moving points along the top line. That's
what this function does. Note the approach has been superceded by an improved
algorithm to generate the unit-sources, which is invoked by passing 
improved_downdip_lines=TRUE to \code{discretized_source_from_source_contours}
}

