% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_script_name.R
\name{parent_script_name}
\alias{parent_script_name}
\title{Find the file used to execute this comment}
\usage{
parent_script_name()
}
\value{
full path to the script from which parent_script_name was called
}
\description{
If the call to \code{parent_script_name} is placed inside a script which is
source'd from R or executed with Rscript, then it should return the full path
to the corresponding file. See discussion on http://stackoverflow.com/questions/1815606/rscript-determine-path-of-the-executing-script
}

