% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{pGR}
\alias{pGR}
\title{Cumulative distribution function for Gutenberg Richter distribution}
\usage{
pGR(q, b, mw_min)
}
\arguments{
\item{q}{vector of quantiles}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}
}
\value{
vector with probability(mw < q | mw >= mw_min)
}
\description{
Note corresponding functions pGR, qGR, dGR, rGR. See ?rGR for an extended
example of fitting with maximum likelihood.
}
\examples{
# Compute exceedance probability for some random Mw values (i.e. conditional
# on the fact that an event occurred)
random_Mw = rGR(10, b=1, mw_min=5)
exceedance_prob = 1 - pGR(random_Mw, b=1, mw_min=5)
}
