% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{plot3d_unit_source_interior_points_cartesian}
\alias{plot3d_unit_source_interior_points_cartesian}
\title{3D rgl plot of the unit source interior points}
\usage{
plot3d_unit_source_interior_points_cartesian(us, aspect = "iso",
  add = FALSE, add_zero_plane = TRUE, ...)
}
\arguments{
\item{us}{List with unit source interior points information: Output of
'unit_source_interior_points_cartesian' or similar}

\item{aspect}{aspect ratio of plot (passed to rgl::plot3d)}

\item{add}{logical Add to an existing plot?}

\item{add_zero_plane}{Draw a plane at depth = 0}

\item{...}{further arguments to plot3d (or points3d if add=TRUE)}
}
\value{
Nothing, but make a plot
}
\description{
3D rgl plot of the unit source interior points
}

