% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{plot_all_tsunami_unit_sources}
\alias{plot_all_tsunami_unit_sources}
\title{Make plot of all tsunami unit sources and their sum}
\usage{
plot_all_tsunami_unit_sources(
  sourcename,
  all_tsunami,
  all_tsunami_rast,
  discrete_source
)
}
\arguments{
\item{sourcename}{name for source zone}

\item{all_tsunami}{list with the tsunami unit sources for all source zones}

\item{all_tsunami_rast}{list of rasters corresponding to all_tsunami}

\item{discrete_source}{discrete source information for sourcename}
}
\value{
Nothing, but make a plot
}
\description{
Make plot of all tsunami unit sources and their sum
}
