% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_events.R
\name{plot_earthquake_event_properties}
\alias{plot_earthquake_event_properties}
\title{Graphical checks of earthquake events}
\usage{
plot_earthquake_event_properties(all_eq_events, relation = "Strasser")
}
\arguments{
\item{all_eq_events}{data.frame with all earthquake events. This could be
generated using the function \code{get_all_earthquake_events}}

\item{relation}{the scaling relation type used for theoretical comparison (see \code{Mw_2_rupture_size})}
}
\value{
nothing, but make a plot
}
\description{
Plot some properties of the earthquake events, with comparisons to scaling
relations (solid lines) and +- 2sd (dashed lines)
}
