% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{plot_unit_source_interior_points_cartesian}
\alias{plot_unit_source_interior_points_cartesian}
\title{Quick plot of a unit source interior points cartesian}
\usage{
plot_unit_source_interior_points_cartesian(us)
}
\arguments{
\item{us}{List with unit source interior points information: Output of
'unit_source_interior_points_cartesian' or similar.}
}
\value{
nothing but make a plot
}
\description{
Quick plot of a unit source interior points cartesian
}
