% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{ptruncGR}
\alias{ptruncGR}
\title{Cumulative distribution function for truncated Gutenberg Richter distribution}
\usage{
ptruncGR(q, b, mw_min, mw_max = Inf)
}
\arguments{
\item{q}{vector of quantiles}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}

\item{mw_max}{Maximum mw}
}
\value{
vector with probability(mw < q | mw >= mw_min)
}
\description{
Note corresponding functions ptruncGR, qtruncGR, dtruncGR, rtruncGR. See '?rtruncGR'
for an extended example using the functions to estimate 'b'.
}
\examples{
# Get the exceedance probability (given that an event occurred) for 10
# random Mw values
random_Mw = rtruncGR(10, b=0.7, mw_min=6.0, mw_max = 9.5)
exceedance_prob = 1 - ptruncGR(random_Mw, b=0.7, mw_min=6.0, mw_max=9.5)

}
