% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{qGR}
\alias{qGR}
\title{Inverse cumulative distribution function for Gutenberg Richter distribution}
\usage{
qGR(p, b, mw_min)
}
\arguments{
\item{p}{vector of probabilities}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}
}
\value{
vector with probability(mw < q | mw >= mw_min)
}
\description{
Note corresponding functions pGR, qGR, dGR, rGR. See ?rGR for an extended example
of fitting with maximum likelihood.
}
\examples{
# Compute the 90th percentile Mw
Mw_90 = qGR(0.9, b=0.8, mw_min=6.0)
}
