% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{qtruncGR}
\alias{qtruncGR}
\title{Inverse cumulative distribution function for truncated Gutenberg Richter distribution}
\usage{
qtruncGR(p, b, mw_min, mw_max = Inf)
}
\arguments{
\item{p}{vector of probabilities}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}

\item{mw_max}{Maximum mw}
}
\value{
vector with Mw values having non-exceedance probabilities p
}
\description{
Note corresponding functions ptruncGR, qtruncGR, dtruncGR, rtruncGR. See ?rtruncGR
for an example of using the functions for b estimation.
}
\examples{
# 90th percentile of truncated GR 
Mw_90 = qtruncGR(0.9, b=1.2, mw_min = 5.5, mw_max = 8.0)

}
