% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_interpolator.R
\name{quad3d_source_interpolator}
\alias{quad3d_source_interpolator}
\title{Interpolate inside a 3d planar source}
\usage{
quad3d_source_interpolator(xy, interpolating_quad)
}
\arguments{
\item{xy}{matrix of xy coordinates ('n' rows, 2 columns)}

\item{interpolating_quad}{Matrix with 4 rows and 3 columns, containing rows
with x,y,depth. It is assumed that interpolating_quad[1:2,1:3] contains the
'left' edge, and interpolating_quad[4:3,1:3] contains the right edge, both 
ordered top to bottom. We further assume that the 'top' and 'bottom' depths
are the same, i.e. \cr
((interpolating_quad[1,3] == interpolating_quad[4,3]) & \cr
 (interpolating_quad[2,3] == interpolating_quad[3,3]))}
}
\value{
xyz coordinates interpolated at xy. Note that x,y should be the same
as the input x,y (we return them anyway for debugging/testing purposes)
}
\description{
Interpolate inside a 3d quadrilateral defined by 4 x,y,depth points.
The 'top' edge and the 'bottom' edge both have constant depth.
}

