% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_mux.R
\name{read_mux2_data_alternative}
\alias{read_mux2_data_alternative}
\title{read a mux2 data file}
\usage{
read_mux2_data_alternative(mux2file, inds = NULL)
}
\arguments{
\item{mux2file}{mux2 filename}

\item{inds}{indices to extract, see \code{read_mux2_data} for details}
}
\value{
see \code{read_mux2_data} for details
}
\description{
This function does basically the same thing as the previous read_mux2_data
file.  It is arguably more elegant. I had hoped it would be faster but it does not seem to be
significantly faster. Anyway it might be useful for debugging so I keep it
here, but am not exporting it into the package namespace
}

