% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_to_ncdf4.R
\name{read_table_from_netcdf}
\alias{read_table_from_netcdf}
\title{Read a netcdf file into a data.frame}
\usage{
read_table_from_netcdf(filename, desired_rows = NULL)
}
\arguments{
\item{filename}{netcdf file}

\item{desired_rows}{integer vector giving the rows to extract. If null, read everything.
Note that rows are read in contiguous chunks (up to 100 at once) for efficiency. For
small datasets it may well be faster to read everything. However, for large datasets
over remote connections this is not possible.}
}
\value{
data.frame with the data
}
\description{
It is assumed that the file was produced by \code{write_table_to_netcdf},
or is compatible with it.
}
