% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_OGR.R
\name{readOGR}
\alias{readOGR}
\title{Read OGR supported file as an sp::SpatialXXXXX object}
\usage{
readOGR(dsn, layer, verbose = TRUE)
}
\arguments{
\item{dsn}{Data source destination (often a folder, or for shapefiles can be "name_of_shapefile.shp")}

\item{layer}{data layer (e.g. for a shapefile "myfile" inside dsn, it is probably "myfile")}

\item{verbose}{}
}
\value{
An object with class of some sp::SpatialXXX object (e.g.
SpatialPolygonsDataFrame, SpatialLinesDataFrame, SpatialPointsDataFrame)
}
\description{
This function was previously provided by rgdal::readOGR. But rgdal is retiring in 2023
so we use as a workaround for the rptha package.
}
