% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffm_fit_simulate_earthquake.R
\name{sffm_events_to_table}
\alias{sffm_events_to_table}
\title{Convert the sffm events information to a table}
\usage{
sffm_events_to_table(all_sffm_events, slip_significant_figures = NULL)
}
\arguments{
\item{all_sffm_events}{output of \code{sffm_make_events_on_discretized_source}}

\item{slip_significant_figures}{integer or NULL. If not NULL, then truncate slip
to this many significant figures when concatenating to string}
}
\value{
data.frame with: columns event_index_string and event_slip_string, having
all the unit-source-indices and their slip values in a character string, separated
by '-' and '_' respectively; and other metadata about the event
}
\description{
Convert the sffm events to a data.frame to ease export to csv or netcdf. Note
the tabular format is not ideal, since the data is relatively unstructured.
For instance, it is necessary to store the indices with non-zero slip in a
character string.  The same issue arises for the slip values themselves.
}
