% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_simulate_earthquake.R
\name{sffm_get_default_model_parameters}
\alias{sffm_get_default_model_parameters}
\title{Get the default sffm model parameters}
\usage{
sffm_get_default_model_parameters()
}
\value{
a list containing default parameters affecting the sffm generation \cr
random_phase_generating_field --  random function f(x) which returns
random numbers of length(x) for the phase generating field \cr
negative_slip_removal_function -- a function f(x) which returns x with negative
values made zero or positive \cr
spatial_slip_decay -- Type of spatial slip decay to apply. Options
are 'gaussian' or 'exponential' or 'none' \cr
recentre_slip -- Logical. Should the location of maximum slip be
adjusted to correspond to the maximum slip location in the input raster? \cr
spectral_amplitude_fun -- Function f(kx, ky, reg_par) to return the spectral amplitude given
kx, ky and a vector of model parameters reg_par. The first two values of
reg_par should  be the corner wavenumbers kcx, kcy IN NUMERICAL SPACE (i.e. the
physical wavenumbers multiplied by dx and dy respectively) \cr
}
\description{
Convenience function to make the sffm_default_model_parameters list. By default
these parameters are used in all sffm_ functions, unless an alternative list
of default parameters is provided. See the help for \code{sffm_simulate} for
an example of using non-default parameters. Default parameters correspond to the
S_{NCF} model of Davies et al., 2015
}
\examples{
sffm_default_par = sffm_get_default_model_parameters()
}
\references{
Davies et al. (2015), 
Tsunami inundation from heterogeneous earthquake slip distributions:
Evaluation of synthetic source models, J. Geophys. Res. Solid Earth, 120,
6431-6451, doi:10.1002/2015JB012272. \cr
}

