% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffm_fit_simulate_earthquake.R
\name{sffm_get_numerical_wavenumbers}
\alias{sffm_get_numerical_wavenumbers}
\title{Compute wavenumbers in numerical space for stochastic slip modelling}
\usage{
sffm_get_numerical_wavenumbers(tg_mat)
}
\arguments{
\item{tg_mat}{= matrix or RasterLayer defining the stochastic slip grid}
}
\value{
list of wavenumbers in numerical space
}
\description{
If tg_mat has dimensions N x M, then the numerical wavenumbers are: \cr
kx' = pmin(0:(M-1), N - (0:(M-1)))/M \cr
ky' = pmin(0:(N-1), N - (0:(N-1)))/N \cr
The function returns kx, ky as matrices with the same dimension as
tg_mat. kx has rows all equal kx', and ky has columns all equal ky'. \cr
These numerical wavenumbers can be adjusted to physical wavenumbers by
division by cellsize dx for kx, or dy for ky.
}

