% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffm_fit_simulate_earthquake.R
\name{sffm_recentre_slip}
\alias{sffm_recentre_slip}
\title{Fix location of peak slip in sffm}
\usage{
sffm_recentre_slip(m1, tg = NULL)
}
\arguments{
\item{m1}{rasterLayer or matrix containing a slip surface to be adjusted}

\item{tg}{rasterLayer or matrix containing a reference slip surface}
}
\value{
recentred version of m1
}
\description{
Move asperities closer to the centre of the rupture,
or to the location of asperities in another slip raster (tg) \cr
If tg is not NULL, then move the max of m1 to the same location as
the max of tg, by re-ordering rows and columns \cr
Otherwise, compute the max along each row -- then re-order the rows so that
the row with the smallest max is at the top of the rupture
(unless is is already at the top or bottom). \cr
Then do the same with the columns, trying to put the smallest max
col on the left (unless it is already at the left or right) \cr
}
