% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_simulate_earthquake.R
\name{sffm_slip_goodness_of_fit}
\alias{sffm_slip_goodness_of_fit}
\title{Goodness of fit for SFFM parameters}
\usage{
sffm_slip_goodness_of_fit(reg_par, tg_rast, verbose = FALSE,
  default_seed = 1, NumRandSf = 200,
  sffm_pars = .sffm_default_model_parameters)
}
\arguments{
\item{reg_par}{= vector of 2 regression parameters  [kcx,kcy in numerical
space]. A 3rd parameter may be accepted in some cases, depending on the
values of reg_par allowed in sffm_pars$spectral_amplitude_function}

\item{tg_rast}{= slip grid raster to fit}

\item{verbose}{= TRUE/FALSE -- Verbose error messages}

\item{NumRandSf}{= Number of slip distributions simulated to compute the
goodness-of-fit of the model}

\item{sffm_pars}{environment containing configuration parameters}

\item{default_seed=}{integer -- passed to set.seed for reproducible fitting
with random fault generation (original .Random.seed is restored at the end)}
}
\value{
A goodness-of-fit measure -- minimising this will lead to the 'best'
        model fit
}
\description{
Given regression parameters, compute a goodness-of-fit statistic of the
model with reg_par and data (tg_mat). This is useful for fitting
statistical models within optimization routines \cr
The statistical model is 'largely' fit in 'numerical' space, not physical
space. \cr
e.g. the smallest non-zero wavenumber resolvable by a DFT is 1/N,
and the nyquist wavenumber is 0.5 \cr
So be careful with units of kcx, kcy etc.
}

