% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{shear_modulus_depth}
\alias{shear_modulus_depth}
\title{Vertical variation of shear modulus with depth}
\usage{
shear_modulus_depth(depth, type = "default")
}
\arguments{
\item{depth}{vector of depths in km. See below for interpretation}

\item{type}{curve type.}
}
\value{
Shear modulus (Pa) at each provided depth value. Options are 'default' and 'prem'
}
\description{
Contains models of the shear modulus (Pascals) as a function of depth (km)
}
\details{
type='default' is a fit by eye to a running median of Bilek and
Lay's (1999) variable rigidity data, with a lower limit of 10GPa for depths
less than 7.5km (reflecting values used to model tsunamis in a few shallow
earthquakes, Geist and Bilek, 2001), and an upper limit of 67 GPa for depths
above 35km (reflecting values from PREM, Dziewonski and Anderson 1981). Note
that 'depth' for this type is interpreted as 'depth below solid earth', because
rptha typically works with such depths.\cr 
type='prem' gives the PREM values from table 4 (second part) of Dziewonski and Anderson 1981.
Note that 'depth' here is 'depth below MSL'.
}
\examples{
   # Make up some depths
   depth_values = c(0, 7.5, 15, 35, 100) 
   # Get the shear modulus at each
   mu = shear_modulus_depth(depth_values, type='default')
   # Note the test below illustrates how mu varies with depth in this 'default' model
   stopifnot(all(abs(mu - c(10, 10, 30, 67, 67)*1e+09) < 1))

   #
   # Compare the default model with Bilek and Lay (1999) data. 
   #
   bl99 = BilekLay99_magnitude_depth_duration()
   #
   # Shear modulus estimate, End-member 1. 
   # Often called the 'constant stress drop' model
   #
   rho = 2.7e+03 # kg/m^3
   L = 10000 # m
   # Note the use of a 'normalised duration' in the paper
   normalised_duration = bl99$Duration * (M0_2_Mw(6.0, inverse=TRUE)/M0_2_Mw(bl99$Mw, inverse=TRUE))**(1/3)
   mu = rho * L^2 / (0.8 * normalised_duration)**2 # (Equation 4 in the paper)

   plot(bl99$Depth, mu, log='y', xlab='Depth (km) ', ylab='Shear modulus (Pa)')
   # Assume bl99 gives depth below MSL, and that ocean depth is ~ 4km
   points(seq(0,100)+4, shear_modulus_depth(seq(0,100)), t='l', col='green')
   # Get a moving median of mu, in 2km depth bins
   depth_bins = round(bl99$Depth/2)
   moving_median = aggregate(mu, list(depth_bins), median)
   points(moving_median[,1]*2, moving_median$x, t='o', col='orange')
   # Common mean value for subduction zones
   abline(h=3e+10, col='black')
   # PREM
   points(seq(1,99, by=0.1), shear_modulus_depth(seq(1,99, by=0.1), type='prem'), t='l', col='blue')

   legend('bottomright', 
       c('Shear modulus (Constant Stress Drop model)', 'Median in 2km depth bins', 
           'Model, with depth adjusted to MSL assuming 4km ocean depth', '30 GPa', 
           'PREM (not designed for subduction interfaces)'), 
       pch=c(1, 1, NA, NA, NA), lty=c(NA, 1, 1, 1, 1), col=c('black', 'orange', 'green', 'black', 'blue'))

}
