% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_scaling.R
\name{slip_from_Mw}
\alias{slip_from_Mw}
\title{Compute slip given Mw (and optionally mu, and a function to compute the area from Mw)}
\usage{
slip_from_Mw(Mw, mu = 3e+10, area_function = function(Mw) {    
  Mw_2_rupture_size(Mw, relation = "Strasser")[1] }, constant = 9.05)
}
\arguments{
\item{Mw}{Moment magnitude}

\item{mu}{Shear Modulus (Pascals)}

\item{area_function}{function which returns the area (in km^2) given Mw}

\item{constant}{The value of constant passed to \code{M0_2_Mw}}
}
\value{
slip in m
}
\description{
This gives an alternative interface to 'slip_from_Mw_area_mu', where a
function to compute area from Mw is provided instead of the area. The default
interface serves as a basic example of a function to compute mean slip given
Mw alone. Functions like this are required in the event_probability calculations
}
\examples{
x = slip_from_Mw(9.0) # Should be roughly 10m
}
