% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{spherical_to_cartesian2d_coordinates}
\alias{spherical_to_cartesian2d_coordinates}
\title{Convert from spherical coordinates to a local 2D cartesian coordinate system}
\usage{
spherical_to_cartesian2d_coordinates(
  coords_lonlat,
  origin_lonlat = NULL,
  r = 6378137
)
}
\arguments{
\item{coords_lonlat}{matrix with 2 columns (lon/lat) in degrees}

\item{origin_lonlat}{vector of length 2 (or matrix with 2 columns and same number of rows as coords_lonlat) 
giving the lon/lat for the origin of
the new coordinate system}

\item{r}{radius of the sphere (earth radius in m as default)}
}
\value{
matrix with x,y in the new coordinate system (units same as r units)
}
\description{
Given lon-lat coordinates for a sphere, and an origin (also lon-lat),
compute the coordinates in a 'local' cartesian 2D coordinate system (x,y):\cr
dlat = (lat - origin_latitude)*degrees_to_radians \cr
dlon = (lon - origin_longitude)*degrees_to_radians \cr
x --> r*dlon*cos(origin_latitude * degrees_to_radians) \cr
y --> r*dlat \cr
This will only have good accuracy for small areas, near to the origin, and
not too close to the poles.
}
\examples{
# Make some random coordinates near to (20, -19)
lonlat = cbind(20 + 0.1*runif(10), -19 + 0.1*runif(10))
origin = c(20, -19)
# Convert to local xy system
new_coords = spherical_to_cartesian2d_coordinates(lonlat, origin)

}
