% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperedGR_moment_release.R
\name{taperedGR_exceedance_rate}
\alias{taperedGR_exceedance_rate}
\title{Tapered Gutenberg Richter exceedance-rate}
\usage{
taperedGR_exceedance_rate(M, N_Mt, Mt, Mc, beta)
}
\arguments{
\item{M}{the moment of interest}

\item{N_Mt}{the number of events with moment exceeding the threshold moment Mt}

\item{Mt}{the threshold moment}

\item{Mc}{the corner moment}

\item{beta}{the beta parameter}
}
\value{
The rate of events with moment > M
}
\description{
Compute the rate of events with Moment > M according to the Tapered Gutenberg
Richter model
}
\examples{
# See example in ?taperedGR_exceedance_rate_derivative

}
