% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperedGR_moment_release.R
\name{taperedGR_exceedance_rate_derivative}
\alias{taperedGR_exceedance_rate_derivative}
\title{Derivative of Tapered Gutenberg Richter exceedance-rate}
\usage{
taperedGR_exceedance_rate_derivative(M, N_Mt, Mt, Mc, beta)
}
\arguments{
\item{M}{the moment of interest}

\item{N_Mt}{the number of events with moment exceeding the threshold moment Mt}

\item{Mt}{the threshold moment}

\item{Mc}{the corner moment}

\item{beta}{the beta parameter}
}
\value{
The derivative
}
\description{
This gives the derivative of the exceedance-rate with respect to the moment,
at the specified moment M
}
\examples{
 # Threshold moment
 Mt = M0_2_Mw(5.5, inverse=TRUE)
 # Corner moment
 Mc = M0_2_Mw(9.2, inverse=TRUE)

 # How many events with M > Mt each year on average?
 N_Mt = 10.47
 # The beta parameter
 beta = 0.54

 #
 # Double check that dN_dM is right by comparison with numerical derivative
 #
 x = M0_2_Mw(7.5, inverse=TRUE)
 # Numerical derivative -- needs a large 'delta h' because x is very large
 dh = x/1e+06 # A suitable numerical derivative increment
 dN_dM_estimate = (taperedGR_exceedance_rate(x+dh, N_Mt, Mt, Mc, beta)-
                   taperedGR_exceedance_rate(x-dh, N_Mt, Mt, Mc, beta)
                  )/(2*dh)
 dN_dM_exact = taperedGR_exceedance_rate_derivative(x, N_Mt, Mt, Mc, beta)
 test = (abs(dN_dM_estimate - dN_dM_exact) < 1e-06*abs(dN_dM_exact))
 stopifnot(test)

 # Quick check of the vectorized version
 dN_dM_check = taperedGR_exceedance_rate_derivative(c(0, x, x, 1), N_Mt, Mt, Mc, beta)
 stopifnot(all(dN_dM_check[c(1, 4)] == 0) & 
           all(dN_dM_check[2:3] == dN_dM_exact))

}
