% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperedGR_moment_release.R
\name{taperedGR_Mc_from_moment_release_rate}
\alias{taperedGR_Mc_from_moment_release_rate}
\title{Corner moment for tapered Gutenberg Richter model}
\usage{
taperedGR_Mc_from_moment_release_rate(L, N_Mt, Mt, beta, simple = FALSE)
}
\arguments{
\item{L}{- specified long-term moment release rate [e.g. from coupled fraction of tectonic convergence]}

\item{N_Mt}{- rate of events with moment exceeding Mt}

\item{Mt}{- threshold moment Mt}

\item{beta}{- taperedGR beta parameter}

\item{simple}{if TRUE then implement the approximation in Kagan (2002),
which is simple and inexact, but quite accurate when Mt << Mc. If FALSE
then solve for Mc using root finding}
}
\value{
corner moment
}
\description{
Given a specified long-term moment release rate and other taperedGR
parameters, compute the corner moment Mc.
}
\examples{
   # Threshold moment
   Mt = M0_2_Mw(5.5, inverse=TRUE)
   # Corner moment
   Mc = M0_2_Mw(9.2, inverse=TRUE)

   # How many events with M > Mt each year on average?
   N_Mt = 10.47
   # The beta parameter
   beta = 0.54

   L = taperedGR_moment_release_rate(N_Mt, Mt, Mc, beta)

   back_computed_Mc = taperedGR_Mc_from_moment_release_rate(L, N_Mt, Mt, beta, simple=FALSE)

   stopifnot(abs(back_computed_Mc - Mc) < Mc*1.0e-07)

   back_computed_Mc_approx = taperedGR_Mc_from_moment_release_rate(L, N_Mt, Mt, beta, simple=TRUE)
   stopifnot(abs(back_computed_Mc - back_computed_Mc_approx) < 0.01*back_computed_Mc)

}
