% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taperedGR_moment_release.R
\name{taperedGR_moment_release_rate}
\alias{taperedGR_moment_release_rate}
\title{Long-term average moment release of tapered Gutenberg Richter model}
\usage{
taperedGR_moment_release_rate(N_Mt, Mt, Mc, beta)
}
\arguments{
\item{N_Mt}{the number of events with moment > threshold moment Mt}

\item{Mt}{the threshold moment}

\item{Mc}{the corner moment}

\item{beta}{the beta parameter}
}
\value{
The long-term moment release rate (moment/year)
}
\description{
This is equal to the integral of [ M *
(-taperedGR_exceedance_rate_derivative) ] dM from Mt to Infinity
}
\examples{
   # Threshold moment
   Mt = M0_2_Mw(5.5, inverse=TRUE)
   # Corner moment
   Mc = M0_2_Mw(9.2, inverse=TRUE)

   # How many events with M > Mt each year on average?
   N_Mt = 10.47
   # The beta parameter
   beta = 0.54

   # Numerically integrate the moment release rate, and compare with the analytical formula
   f<-function(M) - M * taperedGR_exceedance_rate_derivative(M, N_Mt, Mt, Mc, beta)

   # Use "Mc *10000" as an approximation of infinity
   numerical_integral = integrate(f, Mt, Mc*10000, rel.tol=1e-08)

   exact_integral = taperedGR_moment_release_rate(N_Mt, Mt, Mc, beta)

   error = abs(exact_integral - numerical_integral$value)
   stopifnot(error < numerical_integral$value*1e-07)
}
