% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsunami_arrival_time_fast.R
\name{tsunami_maxima_and_arrival_time}
\alias{tsunami_maxima_and_arrival_time}
\title{arrival time calculation}
\usage{
tsunami_maxima_and_arrival_time(
  time,
  stage,
  msl = 0,
  arrival_fraction_of_maxima = 0.2
)
}
\arguments{
\item{time}{vector of times}

\item{stage}{vector of stage values}

\item{msl}{mean sea level. length(msl) == 1. Tsunami size is stage-msl}

\item{arrival_fraction_of_maxima}{We say the tsunami has arrived when (stage-msl) > arrival_fraction_of_maxima*max(stage-msl)}
}
\value{
A vector of length 2 giving max(stage), and the time of arrival.
}
\description{
Interface to the fortran routines
}
\examples{
time = seq(0, 10, len=100)
stage = sin(time)
max_and_arrival = tsunami_maxima_and_arrival_time(time, stage)
stopifnot(max(stage) == max_and_arrival[1])
ind = min(which(time == max_and_arrival[2]))
stopifnot(stage[ind] > 0.2 & stage[ind-1] < 0.2)

}
