% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{tsunami_unit_source_2_raster}
\alias{tsunami_unit_source_2_raster}
\title{Convert the tsunami unit source to a z-displacement raster}
\usage{
tsunami_unit_source_2_raster(tsunami_unit_source, filename = NULL,
  saveonly = FALSE)
}
\description{
Convert the tsunami unit source to a z-displacement raster
}

