% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{unit_source_cartesian_to_okada_tsunami_source}
\alias{unit_source_cartesian_to_okada_tsunami_source}
\title{Convert a cartesian unit source with interior points to an okada tsunami source}
\usage{
unit_source_cartesian_to_okada_tsunami_source(us,
  tsunami_surface_points_cartesian)
}
\arguments{
\item{us}{A unit source with interior points in cartesian coordinates.}

\item{tsunami_surface_points_cartesian}{Points at which to compute the
tsunami deformation in cartesian coordinates}
}
\value{
List with edsp, ndsp, zdsp giving the displacements at the
tsunami_surface_points_cartesian.
}
\description{
Convert a cartesian unit source with interior points to an okada tsunami source
}

