% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_interpolator.R
\name{unit_source_grid_to_SpatialPolygonsDataFrame}
\alias{unit_source_grid_to_SpatialPolygonsDataFrame}
\title{Convert a unit_source_grid to SpatialPolygonsDataFrame}
\usage{
unit_source_grid_to_SpatialPolygonsDataFrame(unit_source_grid)
}
\arguments{
\item{unit_source_grid}{3d array defining the outline of unit sources
on some source-zone contours. Typically obtained as the unit_source_grid
entry from \code{discretized_source_from_source_contours}.}
}
\value{
SpatialPolygonsDataFrame, with attributes giving the down-dip
and along-strike indices of each polygon.
}
\description{
Useful to exploit vector operations in sp/rgeos etc
}
