% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{unit_source_interior_points_cartesian}
\alias{unit_source_interior_points_cartesian}
\title{Fill a unit source with "sub-unit-source" points (for tsunami source integration)}
\usage{
unit_source_interior_points_cartesian(discretized_source, unit_source_index,
  origin = NULL, r = 6378137, approx_dx = NULL, approx_dy = NULL,
  scale_dxdy = 1, depths_in_km = TRUE, plot_source = FALSE)
}
\arguments{
\item{discretized_source}{list containing unit sources information (e.g. output of
discretized_source_from_source_contours)}

\item{unit_source_index}{vector of length 2 with integer index (down-dip,
along-strike) of the unit source to operate on.}

\item{origin}{vector with 2 entries (lon,lat) giving the local cartesian coordinate
system origin for x,y. The depth origin is at 0. If NULL, the lon,lat of the
first coordinate associated with the unit source at unit_source_index is
used.}

\item{r}{Radius of the earth (default gives radius in m)}

\item{approx_dx}{Approximate x spacing of grid points inside the unit source
(same units as r). If NULL, then half the min depth of the unit source is used.}

\item{approx_dy}{Approximate y spacing of grid points inside the unit source
(same units as r). If NULL, then half the min depth of the unit source is used.}

\item{scale_dxdy}{approx_dx and approx_dy are scaled by this constant.
Most useful in the case that the latter are set to NULL since it allows
rescaling of defaults.}

\item{depths_in_km}{logical. If TRUE, input depths are assumed in km, and
are transformed to m (so output depths are in m). Otherwise input depths are
assumed in m}

\item{plot_source}{logical. make a simple plot?}
}
\value{
list containing: 'unit_source': matrix with unit_source_coords in the
cartesian coordinate system + an additional depth column (m); 'origin' the
lon/lat origin of the local coordinate system; r (input argument); 'dx, dy'
the actual dx, dy of the grid points (not exactly the same as the approximate
values); 'grid_points' grid points inside the source in the cartesian
coordinate system; 'grid_point_polygons' a polygon associated with the grid points
}
\description{
Make a locally linear coordinate system for a unit source and compute a
'grid' of points inside the source. These can be used later for tsunami
source integration
}

