% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_to_ncdf4.R
\name{write_table_to_netcdf}
\alias{write_table_to_netcdf}
\title{Write a data.frame to a netcdf file}
\usage{
write_table_to_netcdf(
  dataframe,
  filename,
  global_attributes_list = NULL,
  units = NULL,
  long_names = NULL,
  var_prec = NULL,
  add_session_info_attribute = FALSE,
  force_v4 = FALSE
)
}
\arguments{
\item{dataframe}{the input data.frame to be written to file}

\item{filename}{the output filename}

\item{global_attributes_list}{a named list of global attributes to add to the file, or NULL}

\item{units}{character vector with units for each column of the dataframe, or NULL}

\item{long_names}{character vector with long_name for each column of the dataframe, or NULL}

\item{var_prec}{character vector with data type for each column (possible
values are 'short', 'integer', 'float', 'double', 'char', 'byte'). If NULL it is made automatically.}

\item{add_session_info_attribute}{if true, add a global attribute containing information from sessionInfo()}

\item{force_v4}{logical Do we create a netcdf4 file (TRUE) or not (FALSE, default)}
}
\value{
nothing, but save the file
}
\description{
The unlimited dimension corresponds to the rows of the data.frame, which 
should enable efficient row-wise access to the data.
}
\examples{

 test_table = data.frame(x = c(1,2,3), y=c('a', 'b', 'csdf'), z=c(1.1, 1.2, 1.3))

 write_table_to_netcdf(test_table, 
     file='test.nc', 
     units=c('m', '', 'kg'), 
     long_names=c('head count', 'mychar', 'asdfasdfa'), 
     var_prec=c('float', 'char', 'float'))

 # Clean up
 unlink('test.nc')

}
