% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_OGR.R
\name{writeOGR}
\alias{writeOGR}
\title{Write an sp object to a spatial format}
\usage{
writeOGR(obj, dsn, layer, driver, verbose = FALSE, overwrite_layer = FALSE)
}
\arguments{
\item{obj}{An object of class sf::SpatialXXXXX (e.g. SpatialPolygonsDataFrame)}

\item{dsn}{Output folder for the data}

\item{layer}{Layer name for the data}

\item{driver}{for the output data}

\item{verbose}{print more info}

\item{overwrite_layer}{If FALSE and the file exists, then this will throw an error. Will overwrite if TRUE.}
}
\value{
Invisibly returns the object in sf format (from st_as_sf(obj) )
}
\description{
This function was previously provided by rgdal::writeOGR. But rgdal is retiring in 2023
so we use as a workaround for the rptha package.
}
