% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mux.R
\name{zero_crossing_period}
\alias{zero_crossing_period}
\title{Compute the zero-crossing-period of x}
\usage{
zero_crossing_period(x, dt = 1)
}
\arguments{
\item{x}{numeric vector containing a timeseries with mean (approx) zero}

\item{dt}{The time between consecutive samples of x}
}
\value{
The zero crossing period of x
}
\description{
This function computes both the up-crossing and down-crossing periods and
returns their average
}
\examples{
x = seq(1,2000, by=2)
period = 30
y = sin(x*2*pi/period)
# Check we can compute this
y_period = zero_crossing_period(y, dt=2)
stopifnot(abs(y_period/period - 1) < 0.1)

}
