% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axpy_local.R
\name{axpy_local}
\alias{axpy_local}
\title{Interface to BLAS axpy}
\usage{
axpy_local(y, a, x)
}
\arguments{
\item{y}{matrix or array}

\item{a}{constant of length 1}

\item{x}{matrix or array with same length as y}
}
\value{
Nothing, but updates the value of y in-place
}
\description{
Replaces y with a*x + y. Both x and y must be double and have the same
length. On large arrays, this is faster than using R to do "y = a*x+y"
}
\examples{
y = 1.0*(1:10) # Make double
x = 1.0*(21:30) # Make double
axpy_local(y, 2.0, x)
stopifnot(all(abs(y - (2*(21:30) + 1:10)) < 1.0e-10))

}
