% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{cartesian2d_to_spherical_coordinates}
\alias{cartesian2d_to_spherical_coordinates}
\title{Inverse of spherical_to_cartesian2d_coordinates}
\usage{
cartesian2d_to_spherical_coordinates(coords_xy, origin_lonlat, r = 6378137)
}
\arguments{
\item{coords_xy}{matrix with 2 columns containing x,y coordinates in the local coordinate system}

\item{origin_lonlat}{vector of length 2 with the lon-lat of the origin of the cartesian coordinate system}

\item{r}{radius of the sphere with the same units as x and y (earth radius in m as default)}
}
\value{
matrix with lon/lat of coords_xy
}
\description{
This will only have good accuracy for small areas near to the origin_lonlat
(and not too close to the poles)
}

