% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{compute_grid_point_areas_in_polygon}
\alias{compute_grid_point_areas_in_polygon}
\title{Given a polygon filled with a grid of points, compute an area associated with
each grid point such that the total area = polygon area, and the area is less
for points near the boundary}
\usage{
compute_grid_point_areas_in_polygon(polygon, approx_dx, approx_dy)
}
\arguments{
\item{polygon}{matrix with x,y coordinates defining a polygon}

\item{approx_dx}{approximate x spacing of points}

\item{approx_dy}{approximate y spacing of points}
}
\description{
All grid points MUST be inside the polygon (filter beforehand with
point.in.polygon if required)
}

