% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{discretized_source_approximate_summary_statistics}
\alias{discretized_source_approximate_summary_statistics}
\title{Compute APPROXIMATE summary statistics for all unit sources in a discretized source
zone}
\usage{
discretized_source_approximate_summary_statistics(discretized_source,
  default_rake = 90, default_slip = 1, make_plot = FALSE)
}
\arguments{
\item{discretized_source}{list with an entry 'unit_source_grid' containing a
3 dimensional array defining the vertices of all unit sources for the source
zone (e.g. output of 'discretized_source_from_source_contours')}

\item{default_rake}{numeric degrees, all unit sources are assigned this rake}

\item{default_slip}{numeric m, all unit sources are assigned this slip}

\item{make_plot}{logical, make a plot?}
}
\value{
data.frame with key summary statistics
}
\description{
The quantities that we output follow those in the i-invall format in the old
URSGA code, although some outputs there (e.g. section number) do not apply
with our algorithms
}

