% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{discretized_source_from_source_contours}
\alias{discretized_source_from_source_contours}
\title{Make discretized_source from subduction interface contours}
\usage{
discretized_source_from_source_contours(source_shapefile,
  desired_subfault_length, desired_subfault_width, make_plot = FALSE,
  contour_depth_attribute = "level", contour_depth_in_km = TRUE,
  extend_line_fraction = 0.1, orthogonal_near_trench = FALSE,
  improved_downdip_lines = TRUE)
}
\arguments{
\item{source_shapefile}{character name of line shapefile defining the
subduction interface contours. It should have an attribute giving the depth.
Update: This argument can alternatively be a spatialLinesDataFrame obtained
by reading the file.}

\item{desired_subfault_length}{numeric desired length of subfaults (km)}

\item{desired_subfault_width}{numeric desired width of subfaults (km)}

\item{make_plot}{logical Make a plot?}

\item{contour_depth_attribute}{character The name of the column in the
attribute table giving the contour depth}

\item{contour_depth_in_km}{logical Are contour depths given in km? (If False,
assume 'm')}

\item{extend_line_fraction}{To ensure that contour lines intersect downdip
lines at the left/right edges we extend them by this fraction of the
end-to-end source length. Not required if 'improved_downdip_lines=TRUE'}

\item{orthogonal_near_trench}{move unit source points along the trench to enhance
orthogonality there. Can reduce numerical artefacts at the trench. 
Not required if 'improved_downdip_lines=TRUE'}

\item{improved_downdip_lines}{If TRUE, use an iterative algorithm for computing the downdip
lines (which define unit source boundaries). This should lead to unit-sources
that are more orthogonal. The iterative algorithm can take a few minutes if
there are hundreds of unit-sources in each along-trench direction.}
}
\value{
A list containing: depth_contours The original source contours;
unit_source_grid A 3 dimensional array descrbing the unit source vertices;
discretized_source_dim A vector of length 2 with number-of-sources-down-dip,
number-of-sources-along-strike; fine_downdip_transects A 3 dimensional array
containing densly spaced points along the down-dip transects, which might be
useful for defining sub-unit-source points for tsunami source integration
}
\description{
Given a shapefile with contour lines in lon/lat coordinates defining the
subduction interface (with an attribute defining their depth), partition into
unit sources with chosen approximate length/width. Information on all the
unit sources is held in a 'discretized_source' list
}

