% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_sources.R
\name{discretized_source_summary_statistics}
\alias{discretized_source_summary_statistics}
\title{Compute summary statistics for all unit sources in a discretized source
zone.}
\usage{
discretized_source_summary_statistics(discretized_source, default_rake = 90,
  default_slip = 1, approx_dx = 2000, approx_dy = 2000,
  make_plot = FALSE, depth_in_km = TRUE)
}
\arguments{
\item{discretized_source}{list with an entry 'unit_source_grid' containing a
3 dimensional array defining the vertices of all unit sources for the source
zone (e.g. output of 'discretized_source_from_source_contours')}

\item{default_rake}{numeric degrees, all unit sources are assigned this rake}

\item{default_slip}{numeric m, all unit sources are assigned this slip}

\item{approx_dx}{numeric. Value of approx_dx passed to unit_source_interior_points_cartesian}

\item{approx_dy}{numeric. Value of approx_dy passed to unit_source_interior_points_cartesian}

\item{make_plot}{logical, make a plot?}

\item{depth_in_km}{Are depths in km (TRUE) or meters (FALSE)}
}
\value{
data.frame with key summary statistics
}
\description{
The quantities that we output follow those in the i-invall format in the old
URSGA code, although some outputs there (e.g. section number) do not apply
with our algorithms. The calculations are based on filling the unit sources
with subgrid points (via unit_source_interior_points_cartesian), so should
more accurately account for the interface structure than does the alternative
'approximate summary statistics' routine. The width and length are defined
so that (width x length) = (dipping interface area). \cr
Note that the statistics will not account for edge tapering that may be
applied to the slip (optionally).
}

