% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gBuffer}
\alias{gBuffer}
\title{limited replacement for rgeos::gBuffer using sf functionality}
\usage{
gBuffer(spgeom, width, quadsegs = 5, byid = FALSE)
}
\description{
Like rgeos, this treads all datasets as Cartesian. It can give small
differences in the results of buffered polygons compared to rgeos::gBuffer,
due to the internals of "sf::st_buffer".  While unimportant in most
circumstances, it can change results, e.g.  if it leads to a point near the
boundary being included/excluded. Suggest to use sf::st_buffer for non-legacy applications.
}
