% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rupture_events.R
\name{get_all_earthquake_events}
\alias{get_all_earthquake_events}
\title{Compute all synthetic uniform slip earthquake events for a sourcezone}
\usage{
get_all_earthquake_events(discrete_source = NULL,
  unit_source_statistics = NULL, Mmin = 7.5, Mmax = 9.6, dMw = 0.1,
  mu = 3e+10, constant = 9.05, target_location = NULL)
}
\arguments{
\item{discrete_source}{The discrete source for the source-zone, as from 
\code{discrete_source_from_source_contours}. If NULL, you must provide
unit_source_statistics (prefered approach)}

\item{unit_source_statistics}{Pre-computed unit-source-statistics for the
discrete-source. It is best to provide this, but if it is NULL, a local
computation is made with \code{discretized_source_summary_statistics},
assuming approx_dx=approx_dy=5000, and using default arguments for other
inputs.  It's normally best to compute this yourself beforehand to take
control of those parameters.}

\item{Mmin}{The minimum Mw value in the table}

\item{Mmax}{The maximum Mw value in the table}

\item{dMw}{The increment between the Mw values}

\item{mu}{Shear modulus in Pascals.}

\item{constant}{value of constant passed to \code{M0_2_Mw}}

\item{target_location}{c(longitude, latitude) or NULL. If not NULL, then
only return earthquake events which include the unit source 'closest' to the
target location (computed assuming a spherical earth with lon/lat coordinates)}
}
\value{
A large table containing information on all earthquake events, and
the unit sources they involve
}
\description{
Apply \code{get_all_earthquake_events_of_magnitude_Mw} to a range of Mw
values and combine the results
}

