% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rupture_events.R
\name{get_all_earthquake_events}
\alias{get_all_earthquake_events}
\title{Apply get_all_earthquake_events to a range of Mw values and combine the results}
\usage{
get_all_earthquake_events(discrete_source, unit_source_statistics = NULL,
  Mmin = 7.5, Mmax = 9.6, dMw = 0.1)
}
\arguments{
\item{discrete_source}{The discrete source for the events}

\item{unit_source_statistics}{Pre-computed unit-source-statistics for the
discrete-source. If NULL, a cheap approximation is made (not recommended)}

\item{Mmin}{The minimum Mw value in the table}

\item{Mmax}{The maximum Mw value in the table}

\item{dMw}{The increment between the Mw values}
}
\value{
A large table containing information on all earthquake events, and
the unit sources they involve
}
\description{
Apply get_all_earthquake_events to a range of Mw values and combine the results
}

