% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_rgeos.R
\name{gIntersects}
\alias{gIntersects}
\title{limited replacement for rgeos::gIntersects using sf functionality}
\usage{
gIntersects(spgeom1, spgeom2 = NULL, byid = FALSE, returnDense = TRUE)
}
\description{
limited replacement for rgeos::gIntersects using sf functionality
}
