% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_util.R
\name{lonlat_nearest_neighbours}
\alias{lonlat_nearest_neighbours}
\title{Nearest neighbours on the sphere}
\usage{
lonlat_nearest_neighbours(lonlat1, lonlat2)
}
\arguments{
\item{lonlat1}{2 column matrix with lon,lat coordinates -- we want to find the neigbour
nearest to these points}

\item{lonlat2}{2 column matrix with lon,lat coordiantes. We will return the indices of points
in this matrix which are nearest points in lonlat1}
}
\value{
Vector with length == length(lonlat1[,1]), and value giving an integer index into
lonlat2
}
\description{
For each point in 'lonlat1', find the nearest point on 'lonlat2', assuming
spherical (longitude, latitude) coordinates.
}
\examples{
   # Find points in lonlat2 nearest to points in lonlat1
   lonlat1 = matrix(cbind(c(0,0), c(10, 5), c(-10, 5), c(190, 0)), ncol=2, byrow=TRUE)
   lonlat2 = matrix(cbind(c(0.5,0.5), c(0,5), c(-180,0)), ncol=2, byrow=TRUE)
   index_match = lonlat_nearest_neighbours(lonlat1, lonlat2)
   stopifnot(all(index_match = c(1,2,2,3)))

}
