% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{make_tsunami_unit_source}
\alias{make_tsunami_unit_source}
\title{unit_source water surface deformation}
\usage{
make_tsunami_unit_source(i, j, discrete_source, rake,
  tsunami_surface_points_lonlat, approx_dx = NULL, approx_dy = NULL,
  scale_dxdy = 1, depths_in_km = TRUE, kajiura_smooth = FALSE,
  surface_point_ocean_depths = NULL, kajiura_grid_spacing = NULL,
  kajiura_volume_change_error_threshold = 0.1,
  kajiura_where_deformation_exceeds_threshold = 0,
  minimal_output = FALSE,
  tsunami_function = unit_source_cartesian_to_okada_tsunami_source,
  verbose = FALSE, edge_taper_width = 0, ...)
}
\arguments{
\item{i}{integer. The down-dip index of the unit source}

\item{j}{integer. The along-strike index of the unit source}

\item{discrete_source}{List with discrete source information (e.g. output of
\code{discretized_source_from_source_contours})}

\item{rake}{Rake (degrees) of the slip vector. 0 is along-strike slip, 90 is pure thrust,..}

\item{tsunami_surface_points_lonlat}{matrix with lon/lat coordinates of
points where tsunami initial condition will be evaluated.}

\item{approx_dx}{Approximate x spacing of integration points inside unit
source. See unit_source_interior_points_cartesian.}

\item{approx_dy}{Approximate x spacing of integration points inside unit
source. See unit_source_interior_points_cartesian.}

\item{scale_dxdy}{Adjustment factor for approx_dx and approx_dy}

\item{depths_in_km}{logical. Are input depths in the discrete source in km
(TRUE) or m (FALSE).}

\item{kajiura_smooth}{logical. Should we apply kajiura smoothing to the
z-displacement computed by tsunami_function?}

\item{surface_point_ocean_depths}{numeric vector giving ocean depths (in m)
at the tsunami surface points. Used only for Kajiura filtering}

\item{kajiura_grid_spacing}{Value used for grid_dx and grid_dy in kajiura filter.
If NULL, max(surface_point_ocean_depths)/2 is used}

\item{kajiura_volume_change_error_threshold}{Value of volume_change_error_threshold
passed to kajiura_filter.}

\item{kajiura_where_deformation_exceeds_threshold}{Numeric. The Kajiura
filter is applied to a rectangular region including all points where the
absolute value of the initial deformation (computed with
\code{tsunami_function}) exceeds the given threshold. Use of non-zero values
much smaller than the significant deformation (e.g. 1.0e-04) can lead to a
large reduction in the memory and time required for the Kajiura filter
computation, although if the threshold is too high then the filter might
not be applied to important areas.}

\item{minimal_output}{Logical. If TRUE, set the unit source and
tsunami_surface_points_lonlat to NA in the outputs. These are memory heavy so
in some settings they are best removed.}

\item{tsunami_function}{Function used to make the tunami source from
a cartesian unit source with interior points, the rake, and the surface
points.  Must return a list containing zdsp (a vector of all z displacements
in m), and perhaps also edsp, ndsp.}

\item{verbose}{logical. Print some information on progress}

\item{edge_taper_width}{Distance over which to taper slip on unit sources. 
Values > 0 can lead to less peaked slip on sources with a shallow top edge
depth. Note values>0 also imply some slip occurring outside the unit source
location (due to smoothing), although we ensure seismic moment is conserved}

\item{...}{further arguments to tsunami function}
}
\value{
a list with the cartesian unit source, tsunami source, i,j indices
and tsunami surface points, smoothed deformation, and rake
}
\description{
Function to make a tsunami for the ij unit source in a discrete source.
}
