% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{make_tsunami_unit_source}
\alias{make_tsunami_unit_source}
\title{Function to make a tsunami for the ij unit source in a discrete source.}
\usage{
make_tsunami_unit_source(i, j, discrete_source, rake,
  tsunami_surface_points_lonlat, approx_dx = NULL, approx_dy = NULL,
  scale_dxdy = 1, depths_in_km = TRUE, kajiura_smooth = FALSE,
  surface_point_ocean_depths = NULL, kajiura_grid_spacing = NULL,
  kajiura_volume_change_error_threshold = 0.5, minimal_output = FALSE,
  tsunami_function = unit_source_cartesian_to_okada_tsunami_source, ...)
}
\arguments{
\item{i}{integer. The down-dip index of the unit source}

\item{j}{integer. The along-strike index of the unit source}

\item{discrete_source}{List with discrete source information}

\item{rake}{Rake (degrees) of the slip vector. 0 is along-strike slip, 90 is pure thrust,..}

\item{tsunami_surface_points_lonlat}{matrix with lon/lat coordinates of
points where tsunami initial condition will be evaluated.}

\item{approx_dx}{Approximate x spacing of integration points inside unit
source. See unit_source_interior_points_cartesian.}

\item{approx_dy}{Approximate x spacing of integration points inside unit
source. See unit_source_interior_points_cartesian.}

\item{scale_dxdy}{Adjustment factor for approx_dx and approx_dy}

\item{depths_in_km}{logical. Are input depths in the discrete source in km
(TRUE) or m (FALSE).}

\item{kajiura_smooth}{logical. Should we apply kajiura smoothing to the
z-displacement computed by tsunami_function?}

\item{surface_point_ocean_depths}{numeric vector giving ocean depths (in m)
at the tsunami surface points. Used only for Kajiura filtering}

\item{kajiura_grid_spacing}{Value used for grid_dx and grid_dy in kajiura filter.
If NULL, max(surface_point_ocean_depths)/2 is used}

\item{kajiura_volume_change_error_threshold}{Value of volume_change_error_threshold
passed to kajiura_filter.}

\item{minimal_output}{Logical. If TRUE, set the unit source and
tsunami_surface_points_lonlat to NA in the outputs. These are memory heavy so
in some settings they are best removed.}

\item{tsunami_function}{Function used to make the tunami source from
a cartesian unit source with interior points, the rake, and the surface
points.  Must return a list containing zdsp (a vector of all z displacements
in m), and perhaps also edsp, ndsp.}

\item{...}{further arguments to tsunami function}
}
\value{
a list with the cartesian unit source, tsunami source, i,j indices
and tsunami surface points, smoothed deformation, and rake
}
\description{
Function to make a tsunami for the ij unit source in a discrete source.
}

