% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rupture_scaling.R
\name{Mw_2_rupture_size}
\alias{Mw_2_rupture_size}
\title{Compute earthquake rupture area, width and length from Mw based on an
empirical scaling relation}
\usage{
Mw_2_rupture_size(Mw, relation = "Strasser", detailed = FALSE)
}
\arguments{
\item{Mw}{Moment Magnitude}

\item{relation}{Name for the scaling relation ('Strasser' uses the interface
event relation for Strasser et al 2010.)}

\item{detailed}{logical. If False return a vector with area/width/length,
otherwise provide a list with the latter as well as information on
log10-standard-deviations}
}
\value{
A numeric vector
}
\description{
Units are km and km^2
}

