% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{ptruncGR}
\alias{ptruncGR}
\title{Cumulative distribution function for truncated Gutenberg Richter distribution}
\usage{
ptruncGR(q, b, mw_min, mw_max = Inf)
}
\arguments{
\item{q}{vector of quantiles}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}

\item{mw_max}{Maximum mw}
}
\value{
vector with probability(mw < q | mw >= mw_min)
}
\description{
Note corresponding functions ptruncGR, qtruncGR, dtruncGR, rtruncGR
}
