% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rupture_probabilities.R
\name{rate_of_earthquakes_greater_than_Mw_function}
\alias{rate_of_earthquakes_greater_than_Mw_function}
\title{Get a function to compute the rate of earthquakes with magnitude > Mw}
\usage{
rate_of_earthquakes_greater_than_Mw_function(slip_rate, slip_rate_prob, b,
  b_prob, Mw_min, Mw_min_prob, Mw_max, Mw_max_prob, sourcezone_total_area,
  event_table, event_conditional_probabilities,
  computational_increment = 0.01,
  Mw_frequency_distribution = "truncated_gutenberg_richter")
}
\arguments{
\item{slip_rate}{numeric vector with 1 or more long-term seismogenic slip rates for
the source-zone (m/year). This gives the long-term average slip on the fault
due to seismic events (creep is ignored, so these rates must account for the
rate of coupling)}

\item{slip_rate_prob}{numeric vector with probabilities for the slip_rate parameters}

\item{b}{numeric vector with 1 or more 'b' parameters}

\item{b_prob}{numeric vector with probabilities for the b parameters}

\item{Mw_min}{numeric vector with 1 or more 'Mw_min' parameters}

\item{Mw_min_prob}{numeric vector with probabilities for the Mw_min parameters}

\item{Mw_max}{numeric vector with 1 or more 'Mw_max' parameters}

\item{Mw_max_prob}{numeric vector with probabilities for the Mw_max parameters}

\item{sourcezone_total_area}{The total area of the sourcezone (km^2)}

\item{event_table}{A data.frame containing information on all events in the
source-zone. Output of \code{get_all_earthquake_events}}

\item{event_conditional_probabilities}{A vector of conditional probabilities for
all events in the source-zone (conditional on the fact than an event with
their magnitude did occur). Output of
\code{get_event_probabilities_conditional_on_Mw}}

\item{computational_increment}{For each final branch of the logic tree, the
rate function with those parameters is computed at points in a sequence from
min(Mw_min) to max(Mw_max), with spacing computational_increment (adjusted if
required so that the extremes are included). The function value is then
evaluated via lookup.}

\item{Mw_frequency_distribution.}{Either 'truncated_gutenberg_richter' or
'characteristic_gutenberg_richter'}
}
\value{
function of a numeric vector Mw, which returns the rate of events
with magnitude > Mw
}
\description{
The input parameters are based on a truncated Gutenberg Richter model: \cr
The parameter 'a' is adjusted to account for constant rescaling, and here it
is computed to match the long-term seismogenic slip rate (m/year). Details
are explained below. All other parameters are provided by the user. \cr
To account for uncertainties in those parameters, the user may specify each
as a vector of one-or-more parameters with associated probabilities (which of
course must sum to 1). The code then treats all combinations of those
parameters as having probability equal to the product of the individual
parameter probabilities. It then integrates the associated Gutenberg Richter
models to produce a single Rate function, which is returned. \cr
Next we clarify how the 'a' parameter is constrained to match the long-term seismogenic slip
rate. Given particular values of all the pararameters except a, and a long-term seisomogenic slip
rate (= long_term_slip_rate*coupling_coefficient), the long-term moment rate
should equal the rate integrated from the individual events: \cr
\deqn{ \mu  long_term_seismogenic_slip_rate  sourcezone_area =
\sum_{i \in events} \big{(} \mu event_slip_{i} event_area_{i} Pr(event_{i} | event_of_size_Mw = Mw_{i}) rate_of_events_of_size_Mw_{i}\big{)} }\cr
where Pr(event_{i} | event_of_size_Mw_{i}) gives the conditional probability of each
event with the same Mw in the event table [i.e. if we take all events with a
given Mw from the table, their conditional probabilities should sum to 1].
The meaning of other variables should be clear.
Note that a factor of 10^(a) will appear in the RHS term
rate_of_events_of_size_Mw_{i}, and nowhere else. Hence to compute a, we just
compute the LHS and RHS of the above equations, tentatively assuming a=0 in the latter
case, and then it follows that a = log10(LHS/RHS).
}

