% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_to_ncdf4.R
\name{read_table_from_netcdf}
\alias{read_table_from_netcdf}
\title{Read a netcdf file into a data.frame}
\usage{
read_table_from_netcdf(
  filename,
  desired_rows = NULL,
  chunk_size = 100,
  varnames = NULL
)
}
\arguments{
\item{filename}{netcdf file}

\item{desired_rows}{integer vector giving the rows to extract. If null, read everything.
Note that rows are read in contiguous chunks (up to 100 at once) for efficiency. For
small datasets it may well be faster to read everything. However, for large datasets
over remote connections this is not possible.}

\item{chunk_size}{integer. If desired_rows is not null, then read the file in contiguous
chunks of rows (e.g. 100 rows if chunk_size=100), and discard any undesired rows. In 
the typical case that the desired_rows includes clusters of closely spaced data, the 
use of a large chunk_size is much more efficient than performing a large number of 
separate reads.}

\item{varnames}{Either NULL (in which case all variables are read), or a character vector
giving the names of variables to read.}
}
\value{
data.frame with the data
}
\description{
It is assumed that the file was produced by \code{write_table_to_netcdf},
or is compatible with it.
}
