% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_to_ncdf4.R
\name{read_table_from_netcdf}
\alias{read_table_from_netcdf}
\title{Read a netcdf file into a data.frame}
\usage{
read_table_from_netcdf(filename)
}
\arguments{
\item{filename}{netcdf file}
}
\value{
data.frame with the data
}
\description{
It is assumed that the file was produced by \code{write_table_to_netcdf},
or is compatible with it.
}
