% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutenberg_richter_densities.R
\name{rtruncGR}
\alias{rtruncGR}
\title{Random samples from truncated Gutenberg Richter distribution}
\usage{
rtruncGR(n, b, mw_min, mw_max = Inf)
}
\arguments{
\item{n}{integer giving desired number of samples}

\item{b}{Gutenberg-Richter b value}

\item{mw_min}{Minimum mw}

\item{mw_max}{Maximum mw}
}
\value{
vector with random Mw values
}
\description{
Note corresponding functions ptruncGR, qtruncGR, dtruncGR, rtruncGR
}
