% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffm_fit_simulate_earthquake.R
\name{sffm_make_events_on_discretized_source}
\alias{sffm_make_events_on_discretized_source}
\title{Make synthetic finite fault models on a discretized source}
\usage{
sffm_make_events_on_discretized_source(discretized_source_statistics,
  target_location, target_event_mw, num_events = 1,
  vary_peak_slip_location = TRUE, zero_low_slip_cells_fraction = 0,
  sourcename = "", sffm_sub_sample_size = c(1, 1), mu = 3e+10,
  return_slip_raster = TRUE, uniform_slip = FALSE)
}
\arguments{
\item{discretized_source_statistics}{data.frame returned from e.g. 
\code{discretized_source_summary_statistics} or
\code{discretized_source_approximate_summary_statistics}}

\item{target_location}{numeric vector c(lon,lat) giving the desired peak slip
location. By default the peak slip will be distributed stochastically around this
point within a region of approximately (L/2, W/2), where L, W are the
length/width of uniform-slip earthquakes of this magnitude based on
Strasser's scaling relations. See vary_peak_slip_location}

\item{target_event_mw}{desired magnitude of the earthquake}

\item{num_events}{number of stochastic slip scenarios to produce}

\item{vary_peak_slip_location}{If TRUE, vary the peak slip location in a window
around the target location, as described above. If FALSE, use the same peak
slip location for every synthetic event}

\item{zero_low_slip_cells_fraction}{real between 0 and 1. If > 0,
then we zero all of the smallest slip values, which contribute < 'zero_low_slip_cells_fraction'
to the sum of all slip values on the rupture. A conservative value is 0, 
but small values (e.g. 0.05) might substantially reduce the number of unit-sources
involved in the rupture, with little distortion of the event.}

\item{sourcename}{Name of source (will be included in output list, can be useful for book-keeping)}

\item{sffm_sub_sample_size}{vector of 2 integers. The slip raster is
re-sampled to a higher resolution [sub_sample_size cells in the x/y directions
for each original cell] before generating the sffm, and then re-averaged before
returning the output.}

\item{mu}{Shear modulus (Pascals)}

\item{return_slip_raster}{logical. If TRUE, include the slip as a raster object in
the output list. If FALSE, set the latter to NULL on output.}

\item{uniform_slip}{logical. If TRUE, then the slip will be uniform on each rupture.
This enables the rupture size to vary while still having uniform slip. By default it
is FALSE, and stochastic slip is used.}
}
\value{
A list with length = num_events. Each element of the list is a list
containing the entries slip_matrix, slip_raster, initial_moment, peak_slip_ind,
numerical_corner_wavenumbers, which should be self-explanatory if you are
familiary with \code{sffm_simulate}. Note the slip raster is returned for plotting
convenience only - one limitation is that all raster cells are the same size, whereas in
reality the unit-source sizes vary slightly.
}
\description{
Currently only the S_{NCF} model from Davies et al. (2015) is implemented.
The corner wave-number parameters are generated stochastically based on
the regression relations provided therein. The peak slip location varies
about the target_location, within a region of approximately (L/2, W/2) where
L, W are the length/width of earthquakes of this magnitude according to 
Strasser's scaling relations
}
\examples{

# Get source contours
puysegur = readOGR(system.file('extdata/puysegur.shp', package='rptha'), layer='puysegur')
# Get downdip lines
puysegur_downdip = readOGR(system.file('extdata/puysegur_downdip.shp', package='rptha'), 
    layer='puysegur_downdip')
# Make discretized_source
puysegur_discretized_source = discretized_source_from_source_contours(
    source_shapefile=puysegur,
    desired_subfault_length=50,
    desired_subfault_width=50,
    downdip_lines=puysegur_downdip)
# Get unit source summary statistics
puysegur_summary_statistics = discretized_source_summary_statistics(
    puysegur_discretized_source,
    approx_dx=5000,
    approx_dy=5000)
# Make stochastic slip events, with location/magnitude based on 2009/07/15
# puysegur earthquake
puysegur_sffm_event1 = sffm_make_events_on_discretized_source(
    puysegur_summary_statistics,
    target_location = c(166.568, -45.745),
    target_event_mw = 7.8,
    num_events = 10)

# Plot the slip raster for the first 6 synthetic events
par(mfrow=c(3,2))
for(i in 1:6) plot(puysegur_sffm_event1[[i]]$slip_raster)
}
