% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sffm_fit_simulate_earthquake.R
\name{sffm_make_random_lwkc_function}
\alias{sffm_make_random_lwkc_function}
\title{Create a function to simulate random length/width/kcx/kcy for synthetic
finite fault models.}
\usage{
sffm_make_random_lwkc_function(
  log10kcx_regression_par = c(-0.54, 2.03, 0.22),
  log10kcy_regression_par = c(-0.41, 1.18, 0.19),
  cor_kcx_kcy_residual = 0.68,
  clip_random_parameter_ranges_to_2sd = FALSE,
  relation = "Strasser",
  force_deterministic = FALSE
)
}
\arguments{
\item{log10kcx_regression_par}{vector of length 3 with the gradient,
intercept and residual standard deviation for the regression of log10(kcx) vs
earthquake-magnitude}

\item{log10kcy_regression_par}{vector of length 3 with the gradient,
intercept and residual standard deviation for the regression of log10(kcy) vs
earthquake-magnitude}

\item{cor_kcx_kcy_residual}{Pearson correlation between the log10(kcx) and
log10(kcy) regression residuals}

\item{clip_random_parameter_ranges_to_2sd}{logical. If FALSE, then use the standard
scaling laws. If TRUE, then clip unit-normal random variables to +- 2SD before 
using in scaling laws [thus removing very 'unusual' l/w/kcx/kcy values]}

\item{relation}{scaling relation type passed to \code{Mw_2_rupture_size}}

\item{force_deterministic}{logical Default FALSE. If TRUE, then do not
sample from the predictive uncertainty (i.e. return the same magnitude based
values of L,W,kcx,kcy every time). This is generally undesirable (because
L,W,kcx,kcy show high variability). However, sometimes this level of user control is useful.}
}
\value{
a function f(Mw) which returns a vector or matrix with random
length, width, kcx, kcy values to be used to simulate an sffm with a given magnitude.
Units are km, km, km^{-1}, km^{-1} respectively.
}
\description{
The default parameter values are based on the Strasser et al. (2010)
length/width vs Mw scaling laws, and the kcx/kcy parameters from the S_{NCF}
stochastic slip model of Davies et al (2015). Beware that kcx/kcy are 'physical'
corner wavenumbers (units of distance^{-1}), whereas the function \code{simulate_sffm} requires these
to be transformed to dimensionless 'numerical' corner wavenumbers, by
division by the sffm cell size.
}
\examples{
# Here we use default parameter values
lwkc = sffm_make_random_lwkc_function()
print(lwkc(7.5)) # Random parameters for Mw = 7.5
print(lwkc(7.5)) # More random parameters
print(lwkc(9.2)) # Larger L, W, and smaller kc
# The function can also take a vector, in which case it returns a matrix with one row for each Mw.
print(lwkc(c(7.5, 8.0, 8.5, 9.0)))
# Usage in deterministic mode
lwkc_deterministic = sffm_make_random_lwkc_function(force_deterministic=TRUE)
x1 = lwkc_deterministic(8.0)
x2 = lwkc_deterministic(8.0)
# x1 and x2 should be the same
stopifnot(all(x1 == x2))

}
