% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsunami_sources.R
\name{tsunami_unit_source_2_raster}
\alias{tsunami_unit_source_2_raster}
\title{Convert the tsunami unit source to a z-displacement raster}
\usage{
tsunami_unit_source_2_raster(tsunami_unit_source, filename = NULL,
  saveonly = FALSE)
}
\arguments{
\item{tsunami_unit_source}{output of make_tsunami_unit_source or similar}

\item{filename}{Name for output raster file. If NULL, no file is saved.}

\item{saveonly}{logical. If FALSE, return the raster object as well as
saving. If TRUE but filename is not NULL, then save the file, but return NULL}
}
\value{
Either a raster, or NULL. Can save the raster to a file as a side effect
}
\description{
Convert the tsunami unit source to a z-displacement raster
}

