% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stage_vs_rate_curve_fast.R
\name{weighted_percentile}
\alias{weighted_percentile}
\title{Get a percentile from an empirical distribution}
\usage{
weighted_percentile(vals, weights, p)
}
\arguments{
\item{vals}{vector of values}

\item{weights}{vector of non-negative weights (one for each value). These will be
normalised to weights/sum(weights) inside the function}

\item{p}{vector, with length 1 or more, containing values for 'p' as defined above. We require 0<=p<=1}
}
\value{
A vector 'X' with the same length as 'p', as defined above
}
\description{
Suppose a distribution is defined empirically by a set of values (vals), each
having a weight 'weight' which defines the probability mass function of each vals.
This function returns the smallest value of X in 'vals' such that PR(vals <= X) >= p.
}
\examples{
   vals = c(-10, 5, 3, -4, 5, 6)
   weights = c(0.5, 0.1, 0.1, 0.1, 0.1, 0.1)

   test_p = c(0.0, 0.5, 0.501, 0.6, 0.601, 0.7, 0.701, 0.8, 0.9, 0.901, 1.0)
   empirical_p = weighted_percentile(vals, weights, test_p)

   expected_vals = c(-10, -10, -4, -4, 3, 3, 5, 5, 5, 6, 6)
   stopifnot(all(empirical_p == expected_vals))

}
